% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{acf.signf}
\alias{acf.signf}
\title{Check whether ACF of residuals is significant in any lags.}
\usage{
acf.signf(resid, lag.max = 4, return.p = FALSE)
}
\arguments{
\item{resid}{Vector of residuals.}

\item{lag.max}{Only check from lag 1 until lag.max.}

\item{return.p}{Return the smallest p-value of the calculated lags.}
}
\value{
If any significant lags are present TRUE is returned otherwise FALSE. If return.p is TRUE then the smallest of the calculated p-values is returned instead.
}
\description{
Check whether ACF of residuals is significant in any lags.
}
\details{
This corresponds to plotting the ACF using acf() and checking whether any lags has an acf value above the CI limit.
}

