% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.catchunit}
\alias{add.catchunit}
\title{Add catch unit to label}
\usage{
add.catchunit(lab, cu)
}
\arguments{
\item{lab}{Base label}

\item{cu}{Catch unit as a character string}
}
\value{
Label with added catch unit
}
\description{
Add catch unit to label
}
