% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.col.legend.hor}
\alias{add.col.legend.hor}
\title{Add a legend explaining colors of points (horizontal orientation)}
\usage{
add.col.legend.hor()
}
\value{
Nothing.
}
\description{
Add a legend explaining colors of points (horizontal orientation)
}
