% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.col.legend}
\alias{add.col.legend}
\title{Add a legend indicating point colors of quarters.}
\usage{
add.col.legend(qs, cols, pch = 1)
}
\arguments{
\item{qs}{Quarters to plot legend for.}

\item{cols}{Vector containing colors.}

\item{pch}{Point character.}
}
\value{
Nothing.
}
\description{
Add a legend indicating point colors of quarters.
}

