% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{add.man.scenario}
\alias{add.man.scenario}
\alias{get.TAC,}
\alias{make.man.inp}
\alias{get.TAC}
\title{Define management scenario}
\usage{
add.man.scenario(
  rep,
  scenarioTitle = "",
  maninterval = NULL,
  maneval = NULL,
  ffac = NULL,
  fabs = NULL,
  cfac = NULL,
  cabs = NULL,
  fractiles = list(catch = 0.5, bbmsy = 0.5, ffmsy = 0.5),
  breakpointB = 0,
  limitB = 0,
  safeguardB = list(limitB = 0, prob = 0.95),
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  ctol = 0.001,
  evalBreakpointB = 0,
  evalLimitB = 0,
  verbose = TRUE,
  dbg = 0,
  mancheck = TRUE
)

get.TAC(
  rep,
  scenarioTitle = "",
  maninterval = NULL,
  maneval = NULL,
  ffac = NULL,
  fabs = NULL,
  cfac = NULL,
  cabs = NULL,
  fractiles = list(catch = 0.5, bbmsy = 0.5, ffmsy = 0.5),
  breakpointB = 0,
  limitB = 0,
  safeguardB = list(limitB = 0, prob = 0.95),
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  ctol = 0.001,
  evalBreakpointB = 0,
  evalLimitB = 0,
  verbose = TRUE,
  dbg = 0,
  mancheck = TRUE
)

make.man.inp(
  rep,
  scenarioTitle = "",
  maninterval = NULL,
  maneval = NULL,
  ffac = NULL,
  fabs = NULL,
  cfac = NULL,
  cabs = NULL,
  fractiles = list(catch = 0.5, bbmsy = 0.5, ffmsy = 0.5),
  breakpointB = 0,
  limitB = 0,
  safeguardB = list(limitB = 0, prob = 0.95),
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  ctol = 0.001,
  evalBreakpointB = 0,
  evalLimitB = 0,
  verbose = TRUE,
  dbg = 0,
  mancheck = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{scenarioTitle}{Title of scenario (default: \code{'customScenario_X'},
where X is an integer equal to the number of scenarios with the same name
in \code{rep$man} plus 1, e.g. \code{'customScenario_3'}).}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: \code{maninterval = c(2020.25,2021.25)}.
Default: NULL.}

\item{maneval}{Time at which to evaluate model states. Example: \code{maneval
= 2021.25}. Default: NULL.}

\item{ffac}{Factor to multiply current fishing mortality by (default: NULL).}

\item{fabs}{Absolute fishing mortality for management period (default: NULL).}

\item{cfac}{Factor to multiply current catch by (default: NULL). Please refer
to the details for more information.}

\item{cabs}{Absolute catch for the management period (default: NULL).}

\item{fractiles}{List defining the fractiles of the 3 distributions of
'catch', 'bbmsy', and 'ffmsy'. By default (0.5) median is used for all 3
quantities. Please refer to the details for more information.}

\item{breakpointB}{Breakpoints in terms of \eqn{B/B_{MSY}} for the
hockey-stick HCR. By default (0) no breakpoint is assumed. If one value
is provided, F is reduced linearly to zero, if \eqn{B/B_{MSY}} is below
the breakpoint. If two values ara provided, F is reduced linearly to the
lower of the two provided values, if \eqn{B/B_{MSY}} is below the higher
and above the lower value, and F is zero if \eqn{B/B_{MSY}} is below the
lower value. The higher value corresponds to a biomass threshold where F
is reduced and the lower to a limit where F=0. Note that the breakpoints
are evaluated at the start of the management period. Please refer to the
details for more information.}

\item{limitB}{Biomass limit as relative to B_{MSY}, i.e. \eqn{B/B_{MSY}}. By
default (0) no limit is assumed. If \eqn{B/B_{MSY}} is below the limitB
(e.g. 0.3), F is set to 0. Note that the limit is evaluated at the start
of the management period.}

\item{safeguardB}{List defining an optional precautionary buffer by means of
a biomass reference level relative to \eqn{B/B_{MSY}} (\code{'limitB'};
default: 0, i.e. deactivating the PA buffer) and the risk aversion
probability (\code{'prob'}; default: 0.95). Please refer to the details
for more information.}

\item{intermediatePeriodCatch}{Catch during intermediate period, e.g. last
year's TAC (default: \code{NULL}). Please refer to the details for more
information.}

\item{intermediatePeriodCatchSDFac}{Factor for the multiplication of the
standard deviation of the catch during the intermediate period (default:
1). Please refer to the details for more information.}

\item{intermediatePeriodCatchList}{List defining catch in the intermediate
period obtaining the elements 'obsC', 'timeC', and 'dtc' (optional
element 'stdevfacC' which is 1 if not provided). Please refer to the
details for more information.}

\item{ctol}{Tolerance of \code{nlminb} when finding F that leads to provided
target catch (via arguments \code{cfac} or \code{cabs})}

\item{evalBreakpointB}{Time for the evaluation of the hockey-stick component
of the HCR: 0 indicating start of the management period and 1 indicating
the end of the management period (default: 0).}

\item{evalLimitB}{Time for the evaluation of the biomass limit component of
the HCR: 0 indicating start of the management period and 1 indicating the
end of the management period (default: 0).}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}

\item{mancheck}{Should the time-dependent objects in \code{inp} be checked
against the management time and corrected if necessary? (Default: TRUE)}
}
\value{
\code{add.man.scenario} returns the input object \code{rep} with the
specified HCR added to the \code{man} list.
\code{get.TAC} returns the total allowable catch (TAC) based on the
specified scenario.
\code{make.man.inp} returns the updated \code{inp} list based on specified HCR.
}
\description{
Define management scenario
}
\details{

}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)

## Fishing at Fmsy
rep <- add.man.scenario(rep)

## MSY hockey-stick rule
rep <- add.man.scenario(rep, breakpointB = 0.5)

## MSY hockey-stick rule accounting for uncertainty
rep <- add.man.scenario(rep, fractiles = list(catch=0.35, bbmsy=0.35, ffmsy=0.35), breakpointB=0.5)

## MSY hockey-stick rule with linearly decreasing F between 0.3 and 0.5 Bmsy
rep <- add.man.scenario(rep, fractiles = list(catch=0.35), breakpointB = c(0.3, 0.5))

## Now `rep` includes 4 management scenarios

## Get the TAC when fishing mortality is equal to Fmsy
get.TAC(rep)

## Get TAC for the MSY hockey-stick rule (only using Btrigger)
get.TAC(rep, breakpointB = 0.5)

## Get TAC for the MSY hockey-stick rule (with Btrigger and Blim)
get.TAC(rep, breakpointB = c(0.3, 0.5))

## Get the TAC for the HCR accounting for uncertainty
get.TAC(rep, fractiles = list(catch=0.35, bbmsy=0.35, ffmsy=0.35), breakpointB=0.5)

## Get the TAC for the ICES (2025) recommended HCR (as used in WKMSYSPICT)
get.TAC(rep, fractiles = list(catch=0.35), breakpointB = 0.5, limitB = 0.3)

}
\references{
ICES. 2017. Report of the Workshop on the Development of the ICES
approach to providing MSY advice for category 3 and 4 stocks
(WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/
ACOM:47. 53 pp.

ICES. 2018. Report of the Eighth Workshop on the Development of
Quantitative Assessment Methodologies based on LIFE-history traits,
exploitation characteristics, and other relevant parameters for
data-limited stocks (WKLIFE VIII), 8-12 October 2018, Lisbon,
Portugal. ICES CM 2018/ACOM:40. 172 pp.

ICES.2019. Ninth Workshop on the Development of Quantitative Assessment
Methodologies based on LIFE-history traits, exploitation characteristics,
and other relevant parameters for data-limited stocks (WKLIFE IX).
ICES Scientific Reports. 1:77. 131 pp.http://doi.org/10.17895/ices.pub.5550

ICES 2020. Report of the Ninth Workshop on the Development of
Quantitative Assessment Methodologies based on LIFE-history traits,
exploitation characteristics, and other relevant parameters for
data-limited stocks (WKLIFE X), ICES Scientific Reports. 2:98. 72 pp.
http://doi.org/10.17895/ices.pub.5985
}
