% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.manlines}
\alias{add.manlines}
\title{Add lines to plot indicating result of management scenarios.}
\usage{
add.manlines(
  rep,
  par,
  par2 = NULL,
  index.shift = 0,
  plot.legend = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{par}{The name of the parameter to be plotted.}

\item{par2}{If a second parameter should be used as explanatory variable instead of time.}

\item{index.shift}{Shift initial time point by this index.}

\item{plot.legend}{Logical; should the legend be plotted?}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{\dots}{Passed to \code{lines}.}
}
\value{
Nothing
}
\description{
Add lines to plot indicating result of management scenarios.
}
