% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{arrow.line}
\alias{arrow.line}
\title{Draw a line with arrow heads.}
\usage{
arrow.line(x, y, length = 0.25, angle = 30, code = 2,
  col = par("fg"), lty = par("lty"), lwd = par("lwd"), ...)
}
\arguments{
\item{x}{X coordinates.}

\item{y}{Y coordinates.}

\item{length}{See documentation for arrows.}

\item{angle}{See documentation for arrows.}

\item{code}{See documentation for arrows.}

\item{col}{See documentation for arrows.}

\item{lty}{See documentation for arrows.}

\item{lwd}{See documentation for arrows.}

\item{...}{See documentation for arrows.}
}
\value{
Nothing, but an arrow line is added to the current plot.
}
\description{
Draw a line with arrow heads.
}
\details{
Add to an existing plot a continuous line with arrow heads showing the direction between each data point
}
