% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.EBinf}
\alias{calc.EBinf}
\title{Calculate E(Binfinity), i.e. the fished equilibrium.}
\usage{
calc.EBinf(K, n, Fl, Fmsy, sdb2)
}
\arguments{
\item{K}{The carrying capacity.}

\item{n}{Pella-Tomlinson exponent.}

\item{Fl}{Average fishing mortality of the last year.}

\item{Fmsy}{Fishing mortality at MSY.}

\item{sdb2}{Standard deviation squared (variance) of B process.}
}
\value{
E(Binf).
}
\description{
Calculate E(Binfinity), i.e. the fished equilibrium.
}
\details{
If a seasonal pattern in F is imposed the annual average F is used for calculating the expectation. Max() is used to avoid negative values.
}
