% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.gamma}
\alias{calc.gamma}
\title{Calculate gamma from n}
\usage{
calc.gamma(n)
}
\arguments{
\item{n}{Exponent of the Pella-Tomlinson surplus production equation.}
}
\description{
Calculate gamma from n
}
