% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{calc.mase}
\alias{calc.mase}
\title{Calculate Mean Absolute Scaled Error (MASE)}
\usage{
calc.mase(rep, verbose = TRUE)
}
\arguments{
\item{rep}{A valid result from fit.spict}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
A data frame with estimate the MASE and the number of runs used for
    the estimation for each index.
}
\description{
Calculate Mean Absolute Scaled Error (MASE)
}
\details{
A function that calculates the Mean Absolute Scaled Error (MASE) for
    the indices.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep)
calc.mase(rep)
}
