% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.tac}
\alias{calc.tac}
\title{Calculate Total Allowable Catch (TAC)}
\usage{
calc.tac(rep, ffac, fractile_catch = 0.5)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{ffac}{Factor to multiply current F by}

\item{fractile_catch}{The fractile of the catch distribution to be used for
setting the TAC. Default (0.5) corresponds to the median.}
}
\value{
Total Allowable Catch (TAC)
}
\description{
Calculate Total Allowable Catch (TAC)
}
