% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{change.euler}
\alias{change.euler}
\title{Change the Euler time step in the input list}
\usage{
change.euler(inp, dteuler = 1/16)
}
\arguments{
\item{inp}{An input list containing data.}

\item{dteuler}{Euler time step to update the input list with
(default 1/16).}
}
\value{
Input list with updated dteuler.
}
\description{
Change the Euler time step in the input list
}
\details{
The Euler time step (dteuler) affects the length of most
    time related vectors of the input list. This function allows
    the rescaling of all time related vectors of the input list
    based on the specified dteuler.
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
