% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check.euler}
\alias{check.euler}
\title{Check the sensitivity of estimated reference levels to the
    Euler time discretization}
\usage{
check.euler(rep, dteuler = 1/64, nobs = 100, Frange = NULL)
}
\arguments{
\item{rep}{Result of fit.spict().}

\item{dteuler}{Euler time step to check against (default 1/64).}

\item{nobs}{Number of years to simulate (default: 100)}

\item{Frange}{Fishing mortality range to optimise over (2
values). If NULL (default) the range goes from 0.01 to 6.}
}
\value{
Matrix with reference levels with current dteuler ('curr'),
    specified dteuler based on iterative estimation ('alt'), and
    the relative difference between them ('diff').
}
\description{
Check the sensitivity of estimated reference levels to the
    Euler time discretization
}
\details{
The difference between the reference levels is calculated
    by estimating the relative difference ((curr - alt) / alt).
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
