% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{check.man}
\alias{check.man}
\title{Check the consistency of management scenarios in rep}
\usage{
check.man(
  rep,
  maninterval = NULL,
  maneval = NULL,
  verbose = TRUE,
  reportmode0 = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: maninterval = c(2020.25,2021.25). Default:
NULL.}

\item{maneval}{Time at which to evaluate model states. Example: \code{maneval
= 2021.25}. Default: NULL.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{reportmode0}{Should it be checked that the reportmode is 0 (default: TRUE).}
}
\value{
TRUE/FALSE
}
\description{
Check the consistency of management scenarios in rep
}
\details{
Internal function that checks if the fitted spict objects in
    \code{rep$man} have a consistent management interval.
}
