% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{check.man.time}
\alias{check.man.time}
\title{Checks and corrects management time to be within model time}
\usage{
check.man.time(
  x,
  maninterval = NULL,
  maneval = NULL,
  verbose = TRUE,
  printTimeline = FALSE,
  mancheck = TRUE
)
}
\arguments{
\item{x}{Either an input list from \code{check.inp} or a result report as
generated by running \code{fit.spict}.}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: maninterval = c(2020.25,2021.25). Default:
NULL.}

\item{maneval}{Time at which to evaluate model states. Example: maneval =
2021.25. Default: NULL.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{printTimeline}{logical; print the management time line (default: FALSE)}

\item{mancheck}{Should the time-dependent objects in \code{inp} be checked
against the management time and corrected if necessary? (Default: TRUE)}
}
\value{
Updated input list or fitted spict object dependent on type of input.
}
\description{
Checks and corrects management time to be within model time
}
\examples{
data(pol)
inp <- check.inp(pol$albacore)
rep <- fit.spict(inp)

## with an input list
check.man.time(inp)

## with an output list
check.man.time(rep)

}
