% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check.rep}
\alias{check.rep}
\title{Check rep list}
\usage{
check.rep(rep, reportmode0 = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{reportmode0}{Should it be checked that the reportmode is 0 (default: TRUE).}
}
\value{
Nothing
}
\description{
Check rep list
}
\details{
Internal function that checks if rep is fitted spict object.
}
