% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{deactivate.priors}
\alias{deactivate.priors}
\title{Deactivates priors in an input list}
\usage{
deactivate.priors(inp, priors = "all")
}
\arguments{
\item{inp}{An input list containing data.}

\item{priors}{Priors to deactivate. By default ('all') all priors are deactivated.}
}
\value{
Inp list without priors.
}
\description{
Deactivates priors in an input list
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
