% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{extract.hindcast.info}
\alias{extract.hindcast.info}
\title{Extract hindcast info from a fitted spict object}
\usage{
extract.hindcast.info(rep, CI = 0.95, verbose = TRUE)
}
\arguments{
\item{rep}{Result of \code{\link{fit.spict}} that contains hindcasted runs
added by \code{\link{hindcast}}.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90%
confidence intervals. By default (CI = 0.95), the 95% confidence
intervals are estimated.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
A list with hindcast information.
}
\description{
Extract hindcast info from a fitted spict object
}
\details{
Note that a difference in the timing of the index observations of
    less than a month are considered acceptable for the estimation of the
    naive prediction residuals and no warning is printed. If the variability
    exceeds a month, the predictions are still calculated, but a warning is
    printed.
}
