% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{extract.simstats}
\alias{extract.simstats}
\title{Extracts relevant statistics from the estimation of a simulated data set.}
\usage{
extract.simstats(rep, inp = NULL)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{inp}{The input list used as input to the validation.spict function.}
}
\value{
A list containing the relevant statistics.
}
\description{
Extracts relevant statistics from the estimation of a simulated data set.
}
\details{
TBA
}
\examples{
data(pol)
repin <- fit.spict(pol$albacore)
sim <- sim.spict(repin)
rep <- fit.spict(sim)
extract.simstats(rep)
}

