% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fd}
\alias{fd}
\title{Format date}
\usage{
fd(d, dec = 2)
}
\arguments{
\item{d}{Point in time in years as decimal number.}

\item{dec}{Number of decimals.}
}
\value{
Correctly formatted date.
}
\description{
Format date
}
