% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic.R
\name{fit.aspic}
\alias{fit.aspic}
\title{Fits aspic to the data contained in the input file}
\usage{
fit.aspic(input, do.boot = FALSE, nboot = 3000, ciperc = 95,
  verbose = FALSE)
}
\arguments{
\item{input}{A spict input list containing observations.}

\item{do.boot}{Do bootstrap to get uncertainties of estimates?}

\item{nboot}{Number of bootstrap runs (only used if do.boot=TRUE). Prager suggests in the ASPIC manual p. 13 to use nboot > 1000 if ciperc > 80.}

\item{ciperc}{Coverage percentage of bootstrapped confidence intervals.}

\item{verbose}{If TRUE write information to screen.}
}
\value{
List containing aspic results.
}
\details{
Only works on Linux. This furthermore requires that wine is installed and that aspic7 is installed and available to the PATH.
}

