% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meyermillar.R
\name{fit.jags}
\alias{fit.jags}
\title{Fit the Meyer & Millar model using rjags}
\usage{
fit.jags(
  inp,
  fn,
  n.iter = 10000,
  n.chains = 1,
  burnin = round(n.iter/2),
  thin = 1000
)
}
\arguments{
\item{inp}{Input list containing data and settings.}

\item{fn}{Filename of containing BUGS code.}

\item{n.iter}{Number of iterations.}

\item{n.chains}{Number of chains.}

\item{burnin}{Number of burn-in iterations.}

\item{thin}{Thin chains by this value.}
}
\value{
The raw output of rjags::coda.samples.
}
\description{
Fit the Meyer & Millar model using rjags
}
