% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spict.R
\name{fit.spict}
\alias{fit.spict}
\title{Fit a continuous-time surplus production model to data.}
\usage{
fit.spict(inp, dbg = 0)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}

\item{dbg}{Debugging option. Will print out runtime information useful for debugging if set to 1. Will print even more if set to 2.}
}
\value{
A result report containing estimates of model parameters, random effects (biomass and fishing mortality), reference points (Fmsy, Bmsy, MSY) including uncertainties given as standard deviations.
}
\description{
Fit a continuous-time surplus production model to data.
}
\details{
Fits the model using the TMB package and returns a result report containing estimates of model parameters, random effects (biomass and fishing mortality), reference points (Fmsy, Bmsy, MSY) including uncertainties given as standard deviations.

Fixed effects:
\itemize{
  \item{"logm"}{ Log of maximum sustainable yield.}
  \item{"logK"}{ Log of carrying capacity.}
  \item{"logq"}{ Log of catchability vector.}
  \item{"logsdf"}{ Log of standard deviation of fishing mortality process noise.}
  \item{"logsdb"}{ Log of standard deviation of biomass process noise.}
}

Optional parameters (which are normally not estimated):
\itemize{
  \item{"phi"}{ Used when inp$nseasons > 1 to specify the cyclic B spline representing seasonal variation.}
  \item{"logalpha"}{ Proportionality factor for the observation noise of the indices and the biomass process noise: sdi = exp(logalpha)*sdb. (normally set to logalpha=0)}
  \item{"logbeta"}{ Proportionality factor for the observation noise of the catches and the fishing mortality process noise: sdc = exp(logbeta)*sdf. (this is often difficult to estimate and can result in divergence of the optimisation. Normally set to logbeta=0)}
  \item{"logn"}{ Parameter determining the shape of the production curve as in the generalised form of Pella & Tomlinson (1969). Default: log(2).}
}

Random effects:
\itemize{
  \item{"logB"}{ Log of the biomass process given by the continuous-time stochastic Schaefer equation: dB_t = r*B_t*(1-(B_t/K)^n)*dt + sdb*dW_t, where dW_t is Brownian motion.}
  \item{"logF"}{ Log of the fishing mortality process given by: F_t = D_s*G_t, where D_s is a cyclic B spline and dG_t = sigma_F * dV, with dV being Brownian motion.}
}

Derived parameters:
\itemize{
  \item{"logr"}{ Log of intrinsic growth rate (r = 4m/K).}
  \item{"logBmsy"}{ Log of the equilibrium biomass (Bmsy) when fished at Fmsy.}
  \item{"logFmsy"}{ Log of the fishing mortality (Fmsy) leading to the maximum sustainable yield.}
  \item{"MSY"}{ The yield when the biomass is at Bmsy and the fishing mortality is at Fmsy, i.e. the maximum sustainable yield.}
}

The above parameter values can be extracted from the fit.spict() results using get.par().

Model assumptions
\itemize{
  \item{"1"}{The intrinsic growth rate (r) represents a combination of natural mortality, growth, and recruitment.}
  \item{"2"}{The biomass B_t refers to the exploitable part of the stock. Estimates in absolute numbers (K, Bmsy, etc.) should be interpreted in light of this.}
  \item{"3"}{The stock is closed to migration.}
  \item{"4"}{Age and size-distribution are stable in time.}
  \item{"5"}{Constant catchability of the gear used to gather information for the biomass index.}
}
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
summary(rep)
plot(rep)
}

