% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.AIC}
\alias{get.AIC}
\title{Calculate AIC from a rep list.}
\usage{
get.AIC(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
AIC
}
\description{
Calculate AIC from a rep list.
}
