% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.catchindexoverlap}
\alias{get.catchindexoverlap}
\title{Find observations of catch and index that overlap}
\usage{
get.catchindexoverlap(inp)
}
\arguments{
\item{inp}{An input list containing data.}
}
\value{
List containing overlapping catch (y) and index (z) observations and their time vectors.
}
\description{
Find observations of catch and index that overlap
}
