% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.cov}
\alias{get.cov}
\title{Get covariance matrix of two reported quantities not of fixed model parameters. Covariance of fixed model parameters can be found in rep$cov.fixed.}
\usage{
get.cov(rep, parname1, parname2, cor = FALSE)
}
\arguments{
\item{rep}{Result of fit.spict().}

\item{parname1}{Name first parameter.}

\item{parname2}{Name second parameter.}

\item{cor}{If TRUE correlation matrix is reported instead of covariance matrix}
}
\value{
Covariance matrix of specified parameters.
}
\description{
Get covariance matrix of two reported quantities not of fixed model parameters. Covariance of fixed model parameters can be found in rep$cov.fixed.
}
