% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.EBinf}
\alias{get.EBinf}
\title{Calculate E(Binfinity) the fished equilibrium.}
\usage{
get.EBinf(rep)
}
\arguments{
\item{rep}{A result of fit.spict.}
}
\value{
E(Binf).
}
\description{
Calculate E(Binfinity) the fished equilibrium.
}
\details{
If a seasonal pattern in F is imposed the annual average F is used for calculating the expectation.
}
