% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.ffac}
\alias{get.ffac}
\title{Estimate fishing mortality factor minimising risk of overfishing}
\usage{
get.ffac(rep, relstate = "logBpBmsy", frac_relstate = 1,
  problevel = 0.95, reportmode = 1)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{relstate}{Relative state as returned by \code{\link{fit.spict}}
(default = "logBpBmsy").}

\item{frac_relstate}{Fraction of relative state (\code{relstate}; default =
1)}

\item{problevel}{Probability level of risk aversion (default = 0.95).}

\item{reportmode}{Integer between 0 and 2 determining which objects will be
adreported (default = 1).}
}
\value{
Optimised Fishing mortality for P(Bp<Blim)
}
\description{
Estimate fishing mortality factor minimising risk of overfishing
}
