% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.manC}
\alias{get.manC}
\title{Estimate catch for management period based on last catch observations}
\usage{
get.manC(rep, inp)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{inp}{Input list with ffac or catch observation corresponding to
management.}
}
\value{
Table with catches for management period based on last observed
    catches and corresponding times.
}
\description{
Estimate catch for management period based on last catch observations
}
\details{
Internal function that estimates the catch in the management period
    based on the catch observations in the last year. Only catch observations
    in the last year are considered. If the management period is longer than
    a year the catches of the last year are raised. If the management period
    is shorter than a year, but only annual catches are available, the
    respective fraction of the last annual catch observation is used. If both
    the management period and the catch observations are subannual, the
    subannual catches of the respective 'seasons' of the year corresponding
    to the 'season' of the management period are used. Be aware that the
    estimated catch might correspond to a different season(s) than the
    management period both are subannual and some catch observations are
    missing.
}
