% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get.manmax}
\alias{get.manmax}
\title{Get spict object in rep$man with longest time series}
\usage{
get.manmax(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
rep in rep$man which hast the longest time series
}
\description{
Get spict object in rep$man with longest time series
}
