% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get.mfrow}
\alias{get.mfrow}
\title{Get mfrow from the number of plots to be plotted}
\usage{
get.mfrow(n)
}
\arguments{
\item{n}{Number of plots to be plotted.}
}
\value{
Nothing
}
\description{
Get mfrow from the number of plots to be plotted
}
