% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.MP}
\alias{get.MP}
\title{Get the management procedure (harvest control rule, advice
    rule)}
\usage{
get.MP(fractileC = 0.5, fractileFFmsy = 0.5, fractileBBmsy = 0.5,
  pa = FALSE, prob = 0.95, bbmsyfrac = 0.3, stabilityClause = FALSE,
  lower = 0.8, upper = 1.2, dteuler = 1/16, npriorSD = NA, nhist = NA,
  env = globalenv(), package = "dlmtool")
}
\arguments{
\item{fractileC}{The fractile of the catch distribution to be used
for setting the TAC. Default is median (0.5).}

\item{fractileFFmsy}{The fractile of the distribution of
F/Fmsy. Default is 0.5 (median).}

\item{fractileBBmsy}{The fractile of the distribution of
B/Bmsy. Default is 0.5 (median).}

\item{pa}{Logical; indicating if the precautionary approach should
be applied (reduce F if P(B<Blim) < prob). Default is FALSE.}

\item{prob}{Probability for the precautionary approach (see
argument 'pa', default is 0.95).}

\item{bbmsyfrac}{Fraction of B/Bmsy which is defined as threshold
(Blim = 0.3 Bmsy, Btrigger = 0.5 Bmsy; default is 0.3)}

\item{stabilityClause}{Logical; If true TAC is bound between two
values set in lower and upper. Default: FALSE.}

\item{lower}{lower bound of the stability clause. Default is 0.8,
used if stabilityClause = TRUE.}

\item{upper}{upper bound of the stability clause. Default is 1.2,
used if stabilityClause = TRUE.}

\item{npriorSD}{standard deviation of logn prior. By default (NA)
default prior is used}

\item{nhist}{number of historic years to use for assessment
(default = NA, which means to use all available years)}

\item{env}{environment where the harvest control rule function(s)
are assigned to.}

\item{package}{Indicating the package for which to create the MP
for (so far only 'dlmtool' implemented, default).}
}
\value{
A function which can estimate TAC recommendations based on
    SPiCT assessment, taking assessment uncertainty into account.
}
\details{
This function creates harvest control rules (HCRs) which
    can be incorporated into a management strategy evaluation
    framework (DLMtool package). HCRs are saved with a generic name
    to the global environment and the names of the HCRs are
    returned if results of the function are assigned to an
    object. HCR runs a SPiCT assessment using catch and relative
    biomass index observations. Stock status estimates are used to
    set the TAC for the next year. TAC can be based on the
    distribution of predicted catches (percentileC) and/or the
    distribution of the Fmsy reference level (percentileFmsy).
    Additionally, a cap can be applied to account for low biomass
    levels (below Bmsy).  Arguments of returned function are 'x' -
    the position in a data-limited mehods data object, 'Data' - the
    data-limited methods data object (see DLMtool), and 'reps' -
    the number of stochastic samples of the TAC recommendation (not
    used for this HCR).  One or several arguments of the function
    can be provided as vectors to generate several HCRs at once
    (several vectors have to have same length).
}
\examples{
\dontrun{
## Put together an operating model from the available DLM toolkit examples
StockEx <- Herring
FleetEx <- Generic_IncE
ObsEx <- Precise_Unbiased
## Remove changes in life history parameters
StockEx@Mgrad <- c(0,0)
StockEx@Kgrad <- c(0,0)
StockEx@Linfgrad <- c(0,0)
StockEx@Size_area_1 <- c(0.5,0.5)
StockEx@Frac_area_1 <- c(0.5,0.5)
StockEx@Prob_staying <- c(0.5,0.5)
## Set the depletion level 
StockEx@D <- c(0.3, 0.4)
## create Operation Model
OMex <- new("OM", Stock = StockEx, Fleet = FleetEx, 
                Obs = ObsEx)
## Set simulation options
OMex@nsim <- 10
OMex@nyears <- 25
OMex@proyears <- 5
## Get SPiCT HCRs
MPname <- c(get.MP(dteuler = 1),
            get.MP(pa=TRUE, dteuler=1))
## run MSE
MSEex <- runMSE(OMex,
                MPs = MPname,
                timelimit = 150,
                CheckMPs = FALSE)
## example plot of results
Pplot2(MSEex, traj="quant", quants=c(0.2, 0.8))
}

}
