% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.msyvec}
\alias{get.msyvec}
\title{If multiple growth rates (r) are used (e.g. for a seasonal model), return specified reference point for all instances of r.}
\usage{
get.msyvec(inp, msy)
}
\arguments{
\item{inp}{An input list as validated by check.inp().}

\item{msy}{Matrix containing reference point values as given by get.par().}
}
\value{
A list containing reference point estimates with upper and lower CI bounds.
}
\description{
If multiple growth rates (r) are used (e.g. for a seasonal model), return specified reference point for all instances of r.
}
