% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.no.active.priors}
\alias{get.no.active.priors}
\title{Get number of active priors}
\usage{
get.no.active.priors(inp)
}
\arguments{
\item{inp}{An input list containing priors (after call to check.inp and/or fit.spict)}
}
\value{
number of active priors
}
\description{
Get number of active priors
}
