% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.osar.pvals}
\alias{get.osar.pvals}
\title{Check whether ACF of catch and index residuals is significant in any lags.}
\usage{
get.osar.pvals(rep)
}
\arguments{
\item{rep}{Result of fit.spict(), but requires that also residuals have been calculated using calc.osa.resic().}
}
\value{
Vector of p-values of length equal to the number of data series.
}
\description{
Check whether ACF of catch and index residuals is significant in any lags.
}
