% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.PAffac}
\alias{get.PAffac}
\title{Get the fishing mortality corresponding to set probability
    of being above certain biomass threshold}
\usage{
get.PAffac(repin, bbmsyfrac = 0.3, prob = 0.95)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{bbmsyfrac}{Fraction of B/Bmsy which is defined as threshold
(Blim = 0.3 Bmsy, Btrigger = 0.5 Bmsy; default is 0.3)}

\item{prob}{Probability to be above threshold (bbmsyfrac; default
is 0.95)}
}
\value{
Optimised Fishing mortality for P(Bp<Blim)
}
