% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.prod.regimes}
\alias{get.prod.regimes}
\title{Calculate AIC for spict models with different productivity
    regime shifts}
\usage{
get.prod.regimes(inp, years = NULL)
}
\arguments{
\item{inp}{An input list containing data.}

\item{years}{Optional; vector with years for the estimation of the
regime shift. By default (NULL) all years of the input data are
used.}
}
\value{
data frame with years and AIC
}
\description{
Calculate AIC for spict models with different productivity
    regime shifts
}
