% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.TACi}
\alias{get.TACi}
\title{Estimate TAC for set fishing mortality factor}
\usage{
get.TACi(repin, ffac, fractileC = 0.5)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{ffac}{Fishing mortality factor}

\item{fractileC}{The fractile of the catch distribution to be used
for setting the TAC. Default is median (0.5).}
}
\value{
Estimate TAC
}
