% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.version}
\alias{get.version}
\title{Get version of spict including git sha1 version if available.}
\usage{
get.version(pkg = "spict")
}
\arguments{
\item{pkg}{Name of package.}
}
\value{
Package version
}
\description{
Get version of spict including git sha1 version if available.
}
