% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getPAffac}
\alias{getPAffac}
\title{Get the fishing mortality to be below certain biomass threshold}
\usage{
getPAffac(repin, bbmsyfrac = 0.5, prob = 0.95, MSEmode = 1)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{bbmsyfrac}{Fraction of B/Bmsy which is defined as threshold}

\item{prob}{Probability to above threshold (bbmsyfrac)}

\item{MSEmode}{logical; indicating if uncertainty should only be estimated for logBlBmsy and logFlFmsy (default).}
}
\value{
g
}
\description{
Get the fishing mortality to be below certain biomass threshold
}
