% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{guess.m2}
\alias{guess.m2}
\title{Use a simple linear regression to guess m (MSY).}
\usage{
guess.m2(obsC, obsI, all.return = FALSE)
}
\arguments{
\item{obsC}{Catch data.}

\item{obsI}{Index data.}

\item{all.return}{If true also return a guess on Emsy (effort at MSY) and components of the linear regression.}
}
\value{
The guess on MSY.
}
\details{
Equations 9.1.7 and 9.1.8 on page 284 of FAO's tropical assessment book are used to guess MSY.
}

