% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{hindcast}
\alias{hindcast}
\title{Conduct hindcasting analysis}
\usage{
hindcast(rep, nyears = 7, reduce.output.size = TRUE, mc.cores = 1)
}
\arguments{
\item{rep}{A valid result from fit.spict.}

\item{nyears}{Number of years of data (catch and effort) to remove (this is
also the total number of model runs).}

\item{reduce.output.size}{logical, if \code{TRUE} (default) hindcasting is
run with \code{getReportCovariance} and \code{getJointPrecision} set as
\code{FALSE}}

\item{mc.cores}{Number of cores for \code{parallel::mclapply} function. By
default 1.}
}
\value{
A spictcls list with the added element 'hindcast' containing the
    results of the hindcasting analysis. Use plotspict.hindcast() to plot these
    results.
}
\description{
Conduct hindcasting analysis
}
\details{
This hindcasting analysis fits spict to the data set while
    sequentially exluding the index observations in the last years.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep, nyears=6)
plotspict.hindcast(rep)
}
