% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{invlogp1}
\alias{invlogp1}
\title{Inverse log "plus one" transform}
\usage{
invlogp1(a)
}
\arguments{
\item{a}{Value to take inverse logp1 of.}
}
\value{
Inverse logp1.
}
\description{
Inverse log "plus one" transform
}
\details{
If a = log(b-1), then the inverse transform is b = 1 + exp(a). Useful for values with lower bound at 1.
}
