% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{latex.figure}
\alias{latex.figure}
\title{Generate latex code for including a figure.}
\usage{
latex.figure(figfile, reportfile, caption = "")
}
\arguments{
\item{figfile}{Path to figure file.}

\item{reportfile}{Path to report file.}

\item{caption}{This character string will be included as the figure caption.}
}
\value{
Nothing.
}
\description{
Generate latex code for including a figure.
}
