% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinp.R
\name{list.possible.priors}
\alias{list.possible.priors}
\title{List parameters to which priors can be added}
\usage{
list.possible.priors()
}
\value{
Prints parameters to which priors can be added.
}
\description{
List parameters to which priors can be added
}
