% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spict.R
\name{make.datin}
\alias{make.datin}
\title{Create data list used as input to TMB::MakeADFun.}
\usage{
make.datin(inp, dbg = 0)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}

\item{dbg}{Debugging option. Will print out runtime information useful for debugging if set to 1.}
}
\value{
List to be used as data input to TMB::MakeADFun.
}
\description{
Create data list used as input to TMB::MakeADFun.
}
