% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make.ellipse}
\alias{make.ellipse}
\title{Calculate confidence ellipsis.}
\usage{
make.ellipse(inds, rep)
}
\arguments{
\item{inds}{Indices of the two reported model parameters.}

\item{rep}{A result report as generated by running fit.spict.}
}
\value{
A matrix with two columns containing the x and y coordinates of the ellipsis.
}
\description{
Calculate confidence ellipsis.
}
\details{
Calculates the confidence ellipsis of two reported model parameters. This is particularly useful as a detailed view of the uncertainty of two correlated parameters.
}
