% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make.fconvec}
\alias{make.fconvec}
\title{Make fcon vector}
\usage{
make.fconvec(inp, fcon)
}
\arguments{
\item{inp}{Input list}

\item{fcon}{Constant to add to F}
}
\value{
Input list containing fconvec
}
\description{
Make fcon vector
}
