% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make.ffacvec}
\alias{make.ffacvec}
\title{Make ffac vector}
\usage{
make.ffacvec(inp, ffac)
}
\arguments{
\item{inp}{Input list}

\item{ffac}{Factor to multiply current F by}
}
\value{
Input list containing ffacvec
}
\description{
Make ffac vector
}
