% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{make.man.inp}
\alias{make.man.inp}
\title{Get updated inp list based on specific HCR}
\usage{
make.man.inp(
  rep,
  scenarioTitle = "",
  maninterval = NULL,
  maneval = NULL,
  ffac = NULL,
  fabs = NULL,
  cfac = NULL,
  cabs = NULL,
  fractiles = list(catch = 0.5, bbmsy = 0.5, ffmsy = 0.5),
  breakpointB = 0,
  safeguardB = list(limitB = 0, prob = 0.95),
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  ctol = 0.001,
  evalBreakpointB = 0,
  verbose = TRUE,
  dbg = 0,
  mancheck = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{scenarioTitle}{Title of scenario (default: \code{'customScenario_X'},
where X is an integer equal to the number of scenarios
with the same name in \code{rep$man} plus 1, e.g. \code{'customScenario_3'}).}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: \code{maninterval = c(2020.25,2021.25)}.
Default: NULL.}

\item{maneval}{Time at which to evaluate model states. Example: \code{maneval
= 2021.25}. Default: NULL.}

\item{ffac}{Factor to multiply current fishing mortality by (default: NULL).}

\item{fabs}{Absolute fishing mortality for management period (default: NULL).}

\item{cfac}{Factor to multiply current catch by (default: NULL). Please refer to
the details for more information.}

\item{cabs}{Absolute catch for the management period (default: NULL).}

\item{fractiles}{List defining the fractiles of the 3 distributions of
'catch', 'bbmsy', and 'ffmsy'. By default (0.5) median is used for all 3
quantities. Please refer to the details for more information.}

\item{breakpointB}{Breakpoints in terms of \eqn{B/B_{MSY}} for the
hockey-stick HCR. By default (0) no breakpoint is assumed. If one value
is provided, F is reduced linearly to zero, if \eqn{B/B_{MSY}} is below
the breakpoint. If two values ara provided, F is reduced linearly to the
lower of the two provided values, if \eqn{B/B_{MSY}} is below the higher
and above the lower value, and F is zero if \eqn{B/B_{MSY}} is below the
lower value. The higher value corresponds to ICES's \eqn{B_{trigger}} and
the lower to ICES's \eqn{B_{lim}}. Note that the breakpoints are
evaluated at the start of the management period. Please refer to the
details for more information.}

\item{safeguardB}{List defining an optional precautionary buffer by means of
a biomass reference level relative to \eqn{B/B_{MSY}} (\code{'limitB'};
default: 0, i.e. deactivating the PA buffer) and the risk aversion
probability (\code{'prob'}; default: 0.95). Please refer to the details
for more information.}

\item{intermediatePeriodCatch}{Catch during intermediate period, e.g. last
year's TAC (default: \code{NULL}). Please refer to the details for more
information.}

\item{intermediatePeriodCatchSDFac}{Factor for the multiplication of the
standard deviation of the catch during the intermediate period (default:
1). Please refer to the details for more information.}

\item{intermediatePeriodCatchList}{List defining catch in the intermediate
period obtaining the elements 'obsC', 'timeC', and 'dtc' (optional
element 'stdevfacC' which is 1 if not provided). Please refer to the
details for more information.}

\item{ctol}{Tolerance of \code{optimise} when finding F that leads to
provided target catch (via arguments \code{cfac} or \code{cabs})}

\item{evalBreakpointB}{Time for the evaluation of the hockey-stick component
of the HCR: 0 indicating start of the mangement period and 1 indicating
the end of the management period (default: 0).}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}

\item{mancheck}{Should the time-dependent objects in \code{inp} be checked
against the management time and corrected if necessary? (Default: TRUE)}
}
\value{
The updated inp list based on specified HCR.
}
\description{
Get updated inp list based on specific HCR
}
\details{
Internal function that creates the required input list for the
    specific HCR. For detailed information, please refer to
    \code{\link{add.man.scenario}}.
}
