% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make.report}
\alias{make.report}
\title{Creates a pdf file containing the summary output and result plots}
\usage{
make.report(rep, reporttitle = "", reportfile = "report.tex",
  summaryoutfile = "summaryout.txt", outdir = ".",
  keep.figurefiles = FALSE, keep.txtfiles = FALSE,
  keep.texfiles = FALSE)
}
\arguments{
\item{rep}{A valid result from fit.spict with OSA residuals.}

\item{reporttitle}{This character string will be printed as the first line of the report.}

\item{reportfile}{character, filename (ending in '.tex') of the report LaTeX file.}

\item{summaryoutfile}{character, filename of the summary output file.}

\item{outdir}{character, directory where all output files are going to be saved.}

\item{keep.figurefiles}{If TRUE generated figure files will not be cleaned up.}

\item{keep.txtfiles}{If TRUE generated txt files will not be cleaned up.}

\item{keep.texfiles}{If TRUE generated tex file will not be cleaned up.}
}
\value{
Nothing.
}
\description{
This function probably works probably only in linux with a LaTeX distribution installed (pdflatex).
}
