% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make.rpellipse}
\alias{make.rpellipse}
\title{Calculate confidence ellipsis for reference points.}
\usage{
make.rpellipse(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
A matrix with two columns containing the x and y coordinates of the ellipsis.
}
\description{
Calculate confidence ellipsis for reference points.
}
\details{
Calculates the confidence ellipsis of logBmsy and logFmsy (last if multiple)
}
