% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make.splinemat}
\alias{make.splinemat}
\title{Make a spline design matrix}
\usage{
make.splinemat(nseasons, order, dtfine = 1/100)
}
\arguments{
\item{nseasons}{Number of seasons}

\item{order}{Order of the spline}

\item{dtfine}{Time between points where spline is evaluated}
}
\value{
Spline design matrix.
}
\description{
Make a spline design matrix
}
