% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.select}
\alias{man.select}
\title{Select management scenarios}
\usage{
man.select(rep, scenarios = "all", verbose = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or
\code{add.man.scenario}.}

\item{scenarios}{Selection of scenarios in preferred order. Can be a vector
with the names of the selected scenarios or numbers indicating their
position in \code{rep$man}, e.g. \code{c(6,2)} for the 6th and 2nd
scenario in the \code{rep$man} list. Setting this argument to \code{NULL}
or "none", removes all scenarios from the spict object. By default
(\code{'all'}), all scenarios are selected.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
A fitted spict object wit selected management scenarios in preferred
    order in \code{rep$man}.
}
\description{
Select management scenarios
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
rep <- manage(rep, c(2,4,6))

## based on names
names(rep$man)
rep1 <- man.select(rep, c("curentF","noF"))

## based on indices
length(rep$man)
rep2 <- man.select(rep, c(1,3))

}
