% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.tac}
\alias{man.tac}
\title{Get the TAC for the management scenarios}
\usage{
man.tac(rep, fractileCatch = 0.5, exp = TRUE, verbose = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or
\code{add.man.scenario}.}

\item{fractileCatch}{Fractile of predicted catch distribution. By default
(0.5), the median is being used.}

\item{exp}{Should tac be reported on natural scale (default: TRUE).}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
rep wit selected management scenarios
}
\description{
Get the TAC for the management scenarios
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
rep <- manage(rep, c(3,4,5))

## Median of predicted catch distributions
man.tac(rep)

## 30th percentile of catch distributions
man.tac(rep, fractileCatch = 0.3)

}
