% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.timeline}
\alias{man.timeline}
\title{Print a schematic to the console visualising the management timeline}
\usage{
man.timeline(x, verbose = TRUE, obsonly = FALSE)
}
\arguments{
\item{x}{Either an input list from \code{check.inp} or a result report as
generated by running \code{fit.spict}.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{obsonly}{Display observation period only}
}
\value{
Nothing
}
\description{
Print a schematic to the console visualising the management timeline
}
\examples{
data(pol)
inp <- check.inp(pol$albacore)
inp$maninterval <- c(1991,1992)
rep <- fit.spict(inp)

## based on an input list
man.timeline(inp)

## based on an output list
man.timeline(rep)

}
