% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manage.R
\name{manage}
\alias{manage}
\title{Calculate predictions under different management scenarios}
\usage{
manage(repin, scenarios = "all", dbg = 0)
}
\arguments{
\item{repin}{Result list from fit.spict().}

\item{scenarios}{Vector of integers specifying which scenarios to run. Default: 'all'.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}
}
\value{
List containing results of management calculations.
}
\description{
Calculate predictions under different management scenarios
}
\details{
Scenarios that are currently implemented include:
\itemize{
  \item{"1"}{ Take a specific catch. Default catch: MSY.}
  \item{"2"}{ Fish at Fmsy.}
  \item{"3"}{ No fishing, reduce to 5\% of last F.}
  \item{"4"}{ Reduce F by X\%. Default X = 25.}
  \item{"5"}{ Increase F by X\%. Default X = 25.}
}
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
repman <- manage(rep)
}

