% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{manbase.select}
\alias{manbase.select}
\title{Change the management base scenario}
\usage{
manbase.select(rep, scenario)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or \code{man.scenario}.}

\item{scenario}{New manbase scenario}
}
\value{
rep wit selected management scenario as the manbase scenario
}
\description{
Change the management base scenario
}
