% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manage.R
\name{mansummary}
\alias{mansummary}
\title{Print management summary.}
\usage{
mansummary(rep, ypred = 1)
}
\arguments{
\item{rep}{Result list as output from manage().}

\item{ypred}{Show results for ypred years into the future.}
}
\value{
Data frame containing management summary.
}
\description{
Print management summary.
}

