% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{meanvar2shaperate}
\alias{meanvar2shaperate}
\title{Convert mean and variance to shape and rate of gamma distribution}
\usage{
meanvar2shaperate(mean, var)
}
\arguments{
\item{mean}{Mean value.}

\item{var}{Variance.}
}
\value{
Vector containing shape and rate parameters.
}
\description{
Convert mean and variance to shape and rate of gamma distribution
}
