% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{modefrac2shaperate}
\alias{modefrac2shaperate}
\title{Convert mode and fractile to shape and rate in Gamma distribution (only for mode>0, i.e. shape >= 1)}
\usage{
modefrac2shaperate(mode, xf, f = 0.9)
}
\arguments{
\item{mode}{x for which f(x) is at the maximum}

\item{xf}{x for which P(X<=x) = f}

\item{f}{fractile}
}
\value{
Vector containing shape and rate parameters.
}
\description{
Convert mode and fractile to shape and rate in Gamma distribution (only for mode>0, i.e. shape >= 1)
}
