% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{mohns_rho}
\alias{mohns_rho}
\title{Calculate Mohn's rho for different estimates}
\usage{
mohns_rho(rep, what = c("FFmsy", "BBmsy"), annualfunc = mean, npeels = NULL)
}
\arguments{
\item{rep}{A valid result from fit.spict}

\item{what}{character vector specifying the quantities}

\item{annualfunc}{function used to convert subannual data into annual}

\item{npeels}{Optional; select less peels than the ones in the fitted object}
}
\value{
A named vector with the Monh's rho value for each quantity.
}
\description{
Calculate Mohn's rho for different estimates
}
\details{
A function that calculates Mohn's rho for selected estimated quantities. The function allows the user to define the method of aggrgating from the subannual time steps (1/dteuler) into annual values; the default is to take the mean.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- retro(rep, nretroyear = 4)
mohns_rho(rep)
}
