% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{osar.acf.plot}
\alias{osar.acf.plot}
\title{Plot osar acf}
\usage{
osar.acf.plot(res, lag.max, pval, ylab)
}
\arguments{
\item{res}{Residuals}

\item{lag.max}{Maximum lag to use in acf calculations.}

\item{pval}{P value}

\item{ylab}{Y-axis label}
}
\value{
Nothing.
}
\description{
Plot osar acf
}
