% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PAnll}
\alias{PAnll}
\title{Get the values of the seasonal spline for m.}
\usage{
PAnll(repin, fac, logBpHat, dbg = 0)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{logBpHat}{Log of aimed for predicted biomass.}

\item{dbg}{Debugging option. Will print out runtime information useful for debugging if set to 1.}

\item{ffac}{Projection are made using F = ffac * F_last.}
}
\value{
Negative log likelihood value for predicted biomass vs. aimed for predicted biomass
}
\description{
Get the values of the seasonal spline for m.
}
