% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.col}
\alias{plot.col}
\title{Plot model points colored depending on the quarter to which they belong.}
\usage{
\method{plot}{col}(time, obs, pch = 1, add = FALSE, typ = "p",
  do.line = TRUE, add.legend = FALSE, ...)
}
\arguments{
\item{time}{Time vector.}

\item{obs}{Observation vector (or residual vector).}

\item{pch}{Point character.}

\item{add}{If TRUE plot is added to the current plot.}

\item{typ}{Plot type.}

\item{...}{Additional plotting arguments.}
}
\value{
Nothing.
}
\description{
Plot model points colored depending on the quarter to which they belong.
}

