% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meyermillar.R
\name{plotmm.priors}
\alias{plotmm.priors}
\alias{plot.priors}
\title{Plot priors of Meyer & Millar model}
\usage{
\method{plot}{priors}(nm, priorsin, add = TRUE, ...)
}
\arguments{
\item{nm}{Name of prior}

\item{priorsin}{List of priors, typically inp$meyermillar$priors.}

\item{add}{If TRUE add to current plot.}

\item{...}{Additional arguments to plot.}
}
\value{
Nothing.
}
\description{
Plot priors of Meyer & Millar model
}
