% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.bbmsy}
\alias{plotspict.bbmsy}
\title{Plot estimated B/Bmsy.}
\usage{
plotspict.bbmsy(rep, logax = FALSE, main = -1, plot.legend = TRUE,
  ylim = NULL, plot.obs = TRUE, qlegend = TRUE, lineat = 1)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}
}
\value{
Nothing.
}
\description{
Plot estimated B/Bmsy.
}
\details{
Plots estimated B/Bmsy.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.bbmsy(rep)
}

