% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.biomass}
\alias{plotspict.biomass}
\title{Plot estimated biomass.}
\usage{
plotspict.biomass(rep, logax = FALSE, main = -1, plot.legend = TRUE,
  ylim = NULL, plot.obs = TRUE, qlegend = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}
}
\value{
Nothing.
}
\description{
Plot estimated biomass.
}
\details{
Plots estimated biomass, Bmsy with confidence limits.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.biomass(rep)
}

