% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.btrend}
\alias{plotspict.btrend}
\title{Plot the expected biomass trend}
\usage{
plotspict.btrend(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
Nothing.
}
\description{
Plot the expected biomass trend
}

