% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.catch.bar}
\alias{plotspict.catch.bar}
\title{Plot observed catch and predictions.}
\usage{
plotspict.catch.bar(
  rep,
  main = "Catch",
  ylim = NULL,
  qlegend = TRUE,
  lcol = "blue",
  xlab = "Time",
  ylab = NULL,
  stamp = get.version()
)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{main}{Title of plot.}

\item{ylim}{Limits for y-axis.}

\item{qlegend}{If TRUE legend explaining colours of observation data is plotted.}

\item{lcol}{Colour of prediction lines.}

\item{xlab}{Label of x-axis.}

\item{ylab}{Label of y-axis.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\description{
Plot observed catch and predictions.
}
\details{
Plots observed catch and predictions using the current F and Fmsy.
    The plot also contains the equilibrium catch if the current F is
    maintained.

If no management scenarios are included in \code{rep$man}, the grey vertical
line corresponds to the time of the last observation. If management scenarios
are included in \code{rep$man}, the prediction and confidence intervals of
the base scenario (\code{rep}) are omitted and instead the projections of the
different management scenarios are drawn in different colours. Dotted lines
of the management scenarios reflect the intermediate period, while solid
lines reflect the management period. Additionally, two vertical lines
correspond to the start and end of the management period.

Be aware that potential catch intervals of more a year, e.g. biennial
assessment so that the intermediate period spans two years, or management
period spans two years, are equally split up into annual intervals.

Be aware of the fact that the catches represent intervals, where the length
of the interval is indicated by \code{dtc}, e.g. with \eqn{dtc = 1, C(1990) =
[1990,1990[}. In the plot the catches (and vertical lines) correspond to the
beginning of the catch interval. It might thus seem as if the time of the
vertical lines and the management interval would not align.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.catch(rep)

}
