% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.ci}
\alias{plotspict.ci}
\title{Plot catch and index data.}
\usage{
plotspict.ci(inp)
}
\arguments{
\item{inp}{An input list containing data.}
}
\value{
Nothing
}
\description{
Plot catch and index data.
}

