% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.data}
\alias{plotspict.data}
\title{Plot input data}
\usage{
plotspict.data(inpin, MSY = NULL, one.index = NULL, stamp = get.version())
}
\arguments{
\item{inpin}{An input list containing data.}

\item{MSY}{Value of MSY.}

\item{one.index}{Integer indicating the number of the index to plot.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing
}
\description{
Plot input data
}

