% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.f}
\alias{plotspict.f}
\title{Plot estimated fishing mortality.}
\usage{
plotspict.f(rep, logax = FALSE, main = -1, plot.legend = TRUE,
  ylim = NULL)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}

\item{main}{Title of plot.}

\item{plot.legend}{If TRUE legend is plotted.}

\item{ylim}{Limits for y-axis.}
}
\value{
Nothing.
}
\description{
Plot estimated fishing mortality.
}
\details{
Plots estimated fishing mortality with Fmsy and associated confidence interval.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.f(rep)
}

