% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.ffmsy}
\alias{plotspict.ffmsy}
\title{Plot estimated relative fishing mortality.}
\usage{
plotspict.ffmsy(rep, logax = FALSE, main = -1, plot.legend = TRUE,
  ylim = NULL)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}
}
\value{
Nothing.
}
\description{
Plot estimated relative fishing mortality.
}
\details{
Plots estimated fishing mortality with Fmsy and associated confidence interval.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.ffmsy(rep)
}

