% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.hcr}
\alias{plotspict.hcr}
\title{Comparison plot of aavailable management scenarios}
\usage{
plotspict.hcr(rep, xlim = c(0, 3), CI = 0.95)
}
\arguments{
\item{rep}{A result report as generated by running \code{\link{fit.spict}}
that contains management scenarios added by \code{\link{manage}} or
\code{\link{add.man.scenario}}.}

\item{xlim}{Numeric vector with the lower and upper x-axis limits}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}
}
\value{
Invisible \code{NULL }
}
\description{
A comparison of all management scenarios in \code{rep$man}. For
    each scenario, the fishing mortality in the management period is shown
    against the relative biomass at the point of the harvest control rule
    (HCR) evaluation.
}
\examples{

rep <- fit.spict(pol$hake)
rep <- manage(rep)
plotspict.hcr(rep)

}
\seealso{
\code{\link{manage}}
}
