% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.hindcast}
\alias{plotspict.hindcast}
\title{Hindcasting plot for indices}
\usage{
plotspict.hindcast(
  rep,
  add.mase = TRUE,
  CI = 0.95,
  verbose = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = "Year",
  ylab = "Index",
  plot.log = TRUE,
  stamp = get.version()
)
}
\arguments{
\item{rep}{A result report as generated by running \code{\link{fit.spict}}
that contains hindcasted runs added by \code{\link{hindcast}}.}

\item{add.mase}{Calculate Mean Absolute Scaled Error (MASE) and add it to the
plots.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90%
confidence intervals. By default (CI = 0.95), the 95% confidence
intervals are estimated.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{xlim}{Limits for x-axis.}

\item{ylim}{Limits for y-axis.}

\item{xlab}{Label for x-axis, "Year" by default.}

\item{ylab}{Label for y-axis, "Index" by default.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
MASE or Invisible \code{NULL}
}
\description{
Hindcasting plot for indices
}
\details{
This function plots the results of the hindcasting cross validation
    analysis for each index.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep, nyears = 5)
plotspict.hindcast(rep)

}
\references{
Carvalho, F., Winker, H., Courtney, D., Kapur, M., Kell, L., Cardinale, M.,
Schirripa, M., Kitakado, T., Yemane, D., Piner, K.R. Maunder, M.N., Taylor,
I., Wetzel, C.R., Doering, K., Johnsonm, K.F., Methot, R. D. (2021). A
cookbook for using model diagnostics in integrated stock assessments.
Fisheries Research, 240, 105959.

Kell, L. T., Kimoto, A., & Kitakado, T. (2016). Evaluation of the prediction
skill of stock assessment using hindcasting. Fisheries research, 183,
119-127.

Kell, L. T., Sharma, R., Kitakado, T., Winker, H., Mosqueira, I., Cardinale,
M., & Fu, D. (2021). Validation of stock assessment methods: is it me or my
model talking?. ICES Journal of Marine Science, 78(6), 2244-2255.

Winker, H., Carvalho, F., & Kapur, M. (2018). JABBA: just another Bayesian
biomass assessment. Fisheries Research, 204, 275-288.
}
\seealso{
\code{\link{hindcast}}
}
