% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.inflsum}
\alias{plotspict.inflsum}
\title{Plots summary of influence statistics of observations.}
\usage{
plotspict.inflsum(rep)
}
\arguments{
\item{rep}{A valid result from calc.influence().}
}
\value{
Nothing.
}
\description{
Plots summary of influence statistics of observations.
}
\details{
TBA
}

