% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.priors}
\alias{plotspict.priors}
\title{Plot priors and posterior distribution.}
\usage{
plotspict.priors(rep, do.plot = 4)
}
\arguments{
\item{rep}{A result from fit.spict.}

\item{do.plot}{Integer defining maximum number of priors to plot.}
}
\value{
Nothing
}
\description{
Plot priors and posterior distribution.
}

