% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.season}
\alias{plotspict.season}
\title{Plot the mean F cycle}
\usage{
plotspict.season(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
Nothing.
}
\description{
Plot the mean F cycle
}
\details{
If seasonal data are available the seasonal cycle in the fishing mortality can be estimated. This function plots this mean F cycle.
}

