% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.seasonProd}
\alias{plotspict.seasonProd}
\title{Plot the mean B cycle}
\usage{
plotspict.seasonProd(rep, stamp = get.version())
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\details{
If seasonal data are available the seasonal cycle in the biomass production can be estimated. This function plots this mean B cycle.
}
