% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.tc}
\alias{plotspict.tc}
\title{Plot time constant.}
\usage{
plotspict.tc(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
Nothing.
}
\description{
Plot time constant.
}
\details{
Plots the time required for the biomass to reach a certain proportion of Bmsy. The time required to reach 95\% of Bmsy is highlighted.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.tc(rep)
}

