% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{pol}
\alias{pol}
\title{Fisheries data included in Polacheck et al. (1993).}
\format{Data are lists containing data and initial values for estimation formatted to be used as an input to fit.spict().}
\source{
Polacheck et al. (1993), Canadian Journal of Fisheries and Aquatic Science, vol 50, pp. 2597-2607.
}
\usage{
data(pol)
}
\description{
Fisheries data included in Polacheck et al. (1993).
}
\details{
Fisheries data for south Atlantic albacore, northern Namibian hake, and New Zealand rock lobster.
}
\examples{
data(pol)
rep <- fit.spict(inp=pol$albacore)
rep <- fit.spict(inp=pol$hake)
rep <- fit.spict(inp=pol$lobster)
}
\keyword{datasets}
