% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{pred.catch}
\alias{pred.catch}
\title{Predict the catch of the prediction interval specified in inp}
\usage{
pred.catch(repin, fac = NULL, get.sd = FALSE, exp = FALSE, dbg = 0)
}
\arguments{
\item{fac}{Factor to multiply current F by. If NULL future will be set to Fmsy.}

\item{get.sd}{Get uncertainty of the predicted catch.}

\item{exp}{If TRUE report exp of log predicted catch.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}

\item{rep}{Result list as output from fit.spict().}
}
\value{
A vector containing predicted catch (possibly with uncertainty).
}

