% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{predict.b}
\alias{predict.b}
\title{Helper function for sim.spict().}
\usage{
\method{predict}{b}(B0, F0, gamma, m, K, n, dt, sdb, btype)
}
\arguments{
\item{B0}{Initial biomass.}

\item{F0}{Fishing mortality.}

\item{gamma}{gamma parameter in Fletcher's Pella-Tomlinson formulation.}

\item{m}{m parameter in Fletcher's Pella-Tomlinson formulation.}

\item{K}{Carrying capacity.}

\item{n}{Pella-Tomlinson exponent.}

\item{dt}{Time step.}

\item{sdb}{Standard deviation of biomass process.}

\item{btype}{If 'lamperti' use Lamperti transformed equation, if 'naive' use naive formulation.}
}
\value{
Predicted biomass at the end of dt.
}
\description{
Helper function for sim.spict().
}
