% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{predict.loge}
\alias{predict.loge}
\title{Helper function for sim.spict().}
\usage{
\method{predict}{loge}(logE0, dt, sdf, efforttype)
}
\arguments{
\item{logE0}{Effort}

\item{dt}{Time step.}

\item{sdf}{Standard deviation of E process.}

\item{efforttype}{If 1 use diffusion on logE, if 2 use diffusion of E with state dependent noise (this induces the drift term -0.5*sdf^2 in log domain)}
}
\value{
Predicted F at the end of dt.
}

