% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{predict.logf}
\alias{predict.logf}
\title{Helper function for sim.spict().}
\usage{
\method{predict}{logf}(logF0, dt, sdf, efforttype)
}
\arguments{
\item{logF0}{Fishing mortality.}

\item{dt}{Time step.}

\item{sdf}{Standard deviation of F process.}

\item{efforttype}{If 1 use diffusion on logF, if 2 use diffusion of F with state dependent noise (this induces the drift term -0.5*sdf^2 in log domain)}
}
\value{
Predicted F at the end of dt.
}
\description{
Helper function for sim.spict().
}
