% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{predict.logmre2}
\alias{predict.logmre2}
\title{Helper function for sim.spict().}
\usage{
\method{predict}{logmre2}(logmre0, dt, sdm, psi, logm)
}
\arguments{
\item{logmre0}{Initial value}

\item{dt}{Time step.}

\item{sdm}{Standard deviation of mre process.}

\item{psi}{Degree of attraction toward mean.}

\item{logm}{Mean logm.}
}
\value{
Predicted mre at the end of dt.
}
