% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{print.man.timeline}
\alias{print.man.timeline}
\title{Print a schematic to the console visualising the management timeline}
\usage{
\method{print}{man.timeline}(x, verbose = TRUE)
}
\arguments{
\item{x}{Either an input list from \code{check.inp} or a result report as
generated by running \code{fit.spict}.}

\item{verbose}{informative text}
}
\value{
Nothing
}
\description{
Print a schematic to the console visualising the management timeline
}
