% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{print.spictcls}
\alias{print.spictcls}
\title{Output a summary of a fit.spict() run.}
\usage{
\method{print}{spictcls}(object)
}
\arguments{
\item{object}{A result report as generated by running fit.spict.}
}
\value{
Nothing.
}
\description{
Output a summary of a fit.spict() run.
}

