% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{probdev}
\alias{probdev}
\title{Estimate deviation between targeted and realised probability of
    overfishing}
\usage{
probdev(ffac, rep, relstate = "logBpBmsy", frac_relstate = 1,
  problevel = 0.95, reportmode = 1, getFrac = FALSE,
  verbose = FALSE)
}
\arguments{
\item{ffac}{Factor to multiply current F by}

\item{rep}{Result list from fit.spict().}

\item{relstate}{Relative state for estimation of probability of overfishing
(one of the relative states returned by \code{\link{fit.spict}}; default
= "logBpBmsy").}

\item{frac_relstate}{Fraction of relative state (\code{relstate}; default =
1)}

\item{problevel}{Probability level of risk aversion (default = 0.95).}

\item{reportmode}{Integer between 0 and 2 determining which objects will be
adreported (default = 1).}

\item{getFrac}{logical; return realised fraction of relative state (default =
FALSE).}

\item{verbose}{logical; print realised fraction of relative state, fishing
mortality factor, and deviation (default = FALSE).}
}
\value{
Returns deviation between targeted and realised probability of being
    above certain biomass threshold under set fishing mortality
}
\description{
Estimate deviation between targeted and realised probability of
    overfishing
}
