% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{process.resid}
\alias{process.resid}
\title{Estimate process residuals}
\usage{
process.resid(rep)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}
}
\value{
Data frame with year, process residuals for biomass and fishing
    mortality.
}
\description{
Estimate process residuals
}
