% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manage.R
\name{prop.F}
\alias{prop.F}
\title{Calculate management for changing F by a given factor.}
\usage{
prop.F(fac, inpin, repin, dbg = 0)
}
\arguments{
\item{fac}{Factor to multiply current F with.}

\item{inpin}{Input list.}

\item{repin}{Results list.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}
}
\value{
List containing results of management calculations.
}
\description{
Calculate management for changing F by a given factor.
}

