% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prune.baserun}
\alias{prune.baserun}
\title{Prune a fitted spict object to the core elements}
\usage{
prune.baserun(rep)
}
\arguments{
\item{rep}{Result of fit.spict().}
}
\value{
Fitted spict core object.
}
\description{
Prune a fitted spict object to the core elements
}
