% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{put.ax}
\alias{put.ax}
\alias{put.xax}
\title{Adds the x-axis to influence plots}
\usage{
put.xax(rep)
}
\arguments{
\item{rep}{A valid result from calc.influence().}
}
\value{
Nothing.
}
\description{
Adds the x-axis to influence plots
}
\details{
TBA
}
