% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic.R
\name{read.aspic.res}
\alias{read.aspic.res}
\title{Reads the parameter estimates of an Aspic result file.}
\usage{
read.aspic.res(filename)
}
\arguments{
\item{filename}{Name of the Aspic result file to read}
}
\value{
Vector containing the parameter estimates.
}
\description{
Reads the parameter estimates of an Aspic result file.
}
\details{
TBA
}
