% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reduce.inp}
\alias{reduce.inp}
\title{Reduce time series to get nested inp lists}
\usage{
reduce.inp(inp, nyears, cutStart = FALSE)
}
\arguments{
\item{inp}{An input list containing data.}

\item{nyears}{number of years to reduce to}

\item{cutStart}{logical; should start or end of time series be cut?}
}
\value{
Reduced inp list
}
\description{
Reduce time series to get nested inp lists
}
