% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{refpointci}
\alias{refpointci}
\title{Draw CI around a reference point using polygon}
\usage{
refpointci(t, ll, ul, cicol = "ivory2")
}
\arguments{
\item{t}{Time vector.}

\item{ll}{Lower limit.}

\item{ul}{Upper limit.}

\item{cicol}{Colour of polygon}
}
\value{
Spline design matrix.
}
\description{
Draw CI around a reference point using polygon
}
