% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{remove.priors}
\alias{remove.priors}
\title{Removes priors from an input list}
\usage{
remove.priors(inp, priors = "all")
}
\arguments{
\item{inp}{An input list containing data.}

\item{priors}{Priors to remove. By default ('all') all priors are removed.}
}
\value{
Inp list without priors.
}
\description{
Removes priors from an input list
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
