% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osar.R
\name{res.diagn}
\alias{res.diagn}
\title{Helper function for calc.osar.resid that calculates residual statistics.}
\usage{
res.diagn(resid, id, name = "")
}
\arguments{
\item{resid}{Residuals from either catches or indices.}

\item{id}{Identifier for residuals e.g. "C".}

\item{name}{Identifier that will be used in warning messages.}
}
\value{
List containing residual statistics in 'diagn', shapiro output in 'shapiro', and bias output in 'bias'.
}
\description{
Helper function for calc.osar.resid that calculates residual statistics.
}
