% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osar.R
\name{res.stats}
\alias{res.stats}
\title{Helper function for calc.osar.resid that calculates residual statistics.}
\usage{
res.stats(resid, name = "")
}
\arguments{
\item{resid}{Residuals from either catches or indices.}

\item{name}{Identifier that will be used in warning messages.}
}
\value{
List containing residual statistics in 'stats', shapiro output in 'shapiro', and bias output in 'bias'.
}
\description{
Helper function for calc.osar.resid that calculates residual statistics.
}

