% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{retape.spict}
\alias{retape.spict}
\title{Retape a fitted spict model based on an updated input list}
\usage{
retape.spict(rep, inp, verbose = FALSE, dbg = 0, mancheck = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{inp}{Input list with updated settings}

\item{verbose}{logical; print informative text (default: FALSE)}

\item{dbg}{Debug flag, dbg=0 no output, dbg=1 some output, dbg=2 more output
(default: 0).}

\item{mancheck}{Logical; Should the time-dependent objects in \code{inp} be
checked against the management time and correct if necessary? (Default: TRUE)}
}
\value{
Retaped spict model
}
\description{
Retape a fitted spict model based on an updated input list
}
\details{
This function reevaluates derived variables of a fitted spict model
    after updating model settings, such as the prediction horizon or fishing
    mortality during the predicted time period.
}
