% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{season.cols}
\alias{season.cols}
\title{Load season colors.}
\usage{
season.cols(modin)
}
\arguments{
\item{modin}{Time vector modulo 1.}
}
\value{
Vector containing season colors.
}
\description{
Load season colors.
}
