% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{set.bref}
\alias{set.bref}
\title{Set Bref}
\usage{
set.bref(rep, indBref = NA)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{indBref}{Index of time to use for estimation of Bref.}
}
\value{
Refitted spictcls object.
}
\description{
Set Bref
}
\details{

}
