% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{shaperate2meanvar}
\alias{shaperate2meanvar}
\title{Convert shape and rate of gamma distribution to mean and variance}
\usage{
shaperate2meanvar(shape, rate)
}
\arguments{
\item{shape}{Shape parameter}

\item{rate}{Rate parameter (scale = 1/rate).}
}
\value{
Vector containing mean and var parameters.
}
\description{
Convert shape and rate of gamma distribution to mean and variance
}
