% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{shorten.inp}
\alias{shorten.inp}
\title{Shorten time series of input data to specified length}
\usage{
shorten.inp(inp, mintime = NULL, maxtime = NULL)
}
\arguments{
\item{inp}{An input list containing data.}

\item{mintime}{Minimum time to include in the modified data
(default: NULL).}

\item{maxtime}{Maximum time to include in the modified data
(default: NULL).}
}
\value{
Inp list with shortened time series.
}
\description{
Shorten time series of input data to specified length
}
\details{
Time is given in decimal notation (e.g. 2005.3). If only
    one of the arguments 'minyear' or 'maxyear' is used, the
    function shortens the time series from either end to the
    respective time.
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
