% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sim.eq}
\alias{sim.eq}
\title{Simulate equilibrium population dynamics based on a spict
    assessment}
\usage{
sim.eq(rep, logF, nobs = 100, opt = 1)
}
\arguments{
\item{rep}{Result of fit.spict().}

\item{logF}{Fishing mortality in log scale.}

\item{nobs}{Number of years to simulate (default: 100)}

\item{opt}{Flag to return average catch over last 5 years (if set
to 1, default), or to return catch and biomass (if set to 2).}
}
\value{
If opt == 1 function returns average catch over last 5
    years of the simulation period, if opt == 2 function returns
    catch and biomass over whole simulation period.
}
\description{
Simulate equilibrium population dynamics based on a spict
    assessment
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
