% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Description.R
\docType{package}
\name{spict}
\alias{spict}
\alias{spict-package}
\title{Fits a continuous-time surplus production model to data}
\description{
Fits a continuous-time surplus production model to data
}
\examples{
rep <- test.spict()
}
\author{
Martin W. Pedersen \email{mawp@dtu.dk}
}
\references{
\url{https://github.com/mawp/spict/}
}
\seealso{
\code{\link{test.spict}}
}
\keyword{assessment}
\keyword{fisheries,}
\keyword{model,}
\keyword{production}

