% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmtool.R
\name{spict2DLMtool}
\alias{spict2DLMtool}
\title{Get function to estimate TAC for the DLMtool package}
\usage{
spict2DLMtool(
  fractileC = 0.5,
  fractileFFmsy = 0.5,
  fractileBBmsy = 0.5,
  uncertaintyCap = FALSE,
  lower = 0.8,
  upper = 1.2,
  env = globalenv()
)
}
\arguments{
\item{fractileC}{The fractile of the catch distribution to be used for setting TAC. Default
is median (0.5).}

\item{fractileFFmsy}{The fractile of the distribution of F/Fmsy. Default is 0.5 (median).}

\item{fractileBBmsy}{The fractile of the distribution of B/Bmsy. Default is 0.5 (median).}

\item{uncertaintyCap}{Logical; If true TAC is bound between two values set in lower and upper. Default: FALSE.}

\item{lower}{lower bound of the uncertainty cap. Default is 0.8, used if uncertaintyCap = TRUE.}

\item{upper}{upper bound of the uncertainty cap. Default is 1.2, used if uncertaintyCap = TRUE.}

\item{env}{environment where the harvest control rule function(s) are assigned to.}
}
\value{
A function which can estimate TAC recommendations based on SPiCT assessment,
  taking assessment uncertainty into account.
}
\description{
This function creates harvest control rules (HCRs) which can be incorporated into a
management strategy evaluation framework (DLMtool package). HCRs are saved with a
generic name to the global environment and name of HCR is returned if results of the
functin are assigned to an object. HCR runs a SPiCT assessment using catch and
relative biomass index observations and stock status estimates are used to set the TAC
for the next year. TAC can be based on the distribution of predicted catches (percentileC)
and/or the distribution of the Fmsy reference level (percentileFmsy).
Additionally, a cap can be applied to account for low biomass levels (below Bmsy).
Arguments of returned function are 'x' - the position in a data-limited mehods data object,
'Data' - the data-limited methods data object (see DLMtool), and 'reps' - the number of
stochastic samples of the TAC recommendation (not used for this HCR).
One or several arguments of the function can be provided as vectors to generate several
HCRs at once (several vectors have to have same length).
}
\examples{
\dontrun{
## Put together an operating model from the available DLM toolkit examples
StockEx <- Herring
FleetEx <- Generic_IncE
ObsEx <- Precise_Unbiased
## Remove changes in life history parameters
StockEx@Mgrad <- c(0,0)
StockEx@Kgrad <- c(0,0)
StockEx@Linfgrad <- c(0,0)
StockEx@Prob_staying <- c(1,1)
## Set the depletion level 
StockEx@D <- c(0.3, 0.4)
## create Operation Model
OMex <- new("OM", Stock = StockEx, Fleet = FleetEx, 
                Obs = ObsEx)
## Set simulation options
OMex@nsim <- 10
OMex@nyears <- 25
OMex@proyears <- 5
## Get SPiCT HCR
MPname <- spict2DLMtool(fractileC=0.3)
## run MSE
MSEex <- runMSE(OMex,
                MPs = MPname,
                timelimit = 150,
                CheckMPs = FALSE)
## example plot of results
Pplot2(MSEex, traj="quant", quants=c(0.2, 0.8))
}

}
