% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stabilityClause}
\alias{stabilityClause}
\title{Apply stability clause to fishing mortality or fishing
    mortality multiplier}
\usage{
stabilityClause(ffac, lo = 0.8, up = 1.2)
}
\arguments{
\item{ffac}{Fishing mortality factor, multiplier or fishing
mortality to apply stability clause to}

\item{lo}{lower boundary for stability clause (default 0.8)}

\item{up}{upper boundary for stability clause (default 1.2)}
}
\value{
Fishing mortality factor after stability clause was applied
}
