% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spictcls}
\alias{summary.spictcls}
\title{Output a summary of a fit.spict() run.}
\usage{
summary.spictcls(object, numdigits = 8)
}
\arguments{
\item{object}{A result report as generated by running fit.spict.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
Nothing.
}
\description{
Output a summary of a fit.spict() run.
}
\details{
The output includes, the convergence message from the optimiser, the likelihood value of the parameters, the parameter estimates with 95% confidence intervals, estimates of derived parameters (BMsy, Fmsy, MSY) with 95% confidence intervals, and predictions of biomass, fishing mortality, and catch for the value of inp$dtpred.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
summary(rep)
}

