% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.diagnostics}
\alias{sumspict.diagnostics}
\title{Diagnostics table}
\usage{
sumspict.diagnostics(rep, ndigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{ndigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing diagnostics information.
}
\description{
Diagnostics table
}
