% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.drefpoints}
\alias{sumspict.drefpoints}
\title{Deternistic reference points of a fit.spict() run.}
\usage{
sumspict.drefpoints(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
data.frame containing deterministic reference points.
}
\description{
Deternistic reference points of a fit.spict() run.
}

