% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.ini}
\alias{sumspict.ini}
\title{Sensitivity to the initial parameter values}
\usage{
sumspict.ini(rep, numdigits)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
list containing diagnostics information.
}
\description{
Sensitivity to the initial parameter values
}
