% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.parest}
\alias{sumspict.parest}
\title{Parameter estimates of a fit.spict() run.}
\usage{
sumspict.parest(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
data.frame containing parameter estimates.
}
\description{
Parameter estimates of a fit.spict() run.
}

