% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.predictions}
\alias{sumspict.predictions}
\title{Predictions of a fit.spict() run.}
\usage{
sumspict.predictions(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
data.frame containing predictions.
}
\description{
Predictions of a fit.spict() run.
}

