% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.priors}
\alias{sumspict.priors}
\title{Fixed parameters table.}
\usage{
sumspict.priors(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing fixed parameter information.
}
\description{
Fixed parameters table.
}

