% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.srefpoints}
\alias{sumspict.srefpoints}
\title{Stochastic reference points of a fit.spict() run.}
\usage{
sumspict.srefpoints(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
data.frame containing stochastic reference points.
}
\description{
Stochastic reference points of a fit.spict() run.
}

