% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.states}
\alias{sumspict.states}
\title{State estimates of a fit.spict() run.}
\usage{
sumspict.states(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
data.frame containing state estimates.
}
\description{
State estimates of a fit.spict() run.
}

