% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.stockpars}
\alias{sumspict.stockpars}
\title{Parameter estimates of a stock in a fit.spict() run.}
\usage{
sumspict.stockpars(rep, stock = 1, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stock}{Number of the stock whose parameters to present.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing parameter estimates.
}

