% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{take.c}
\alias{take.c}
\title{Calculate management when taking a constant catch (proxy for setting a TAC).}
\usage{
take.c(catchfac, inpin, repin, dbg = 0)
}
\arguments{
\item{catchfac}{Take the catch corresponding to the catch at manstart time catchfac.}

\item{inpin}{Input list.}

\item{repin}{Results list.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}
}
\value{
List containing results of management calculations.
}

