% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{test.spict}
\alias{test.spict}
\title{Example of a spict analysis.}
\usage{
test.spict(dataset = "albacore")
}
\arguments{
\item{dataset}{Specify one of the three test data sets: 'albacore', 'hake', 'lobster'. These can be accessed with the command data(pol).}
}
\value{
A result report as given by fit.spict().
}
\description{
Example of a spict analysis.
}
\details{
Loads a data set, fits the model, calculates one-step-ahead residuals, plots the results.
}
\examples{
rep <- test.spict()
}
