% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{trans2real}
\alias{trans2real}
\title{Get real parameter values from transformed ones.}
\usage{
trans2real(vals, nms, chgnms = TRUE)
}
\arguments{
\item{vals}{Parameters in transformed domain.}

\item{nms}{Names of transformed parameters (including log etc.)}

\item{chgnms}{Remove transformation indication from the parameter names (e.g. remove log from logK).}
}
\value{
Parameter values in the natural domain.
}
\description{
Get real parameter values from transformed ones.
}
