% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{txt.stamp}
\alias{txt.stamp}
\title{Add spict version to plot}
\usage{
txt.stamp(string = get.version(), cex = 0.5, do.flag = NULL,
  new.flag = TRUE)
}
\arguments{
\item{string}{Character string to stamp.}

\item{cex}{Stamp cex.}

\item{do.flag}{If NULL stamp will be added if not in a multi plot,
i.e. mean(par()$mfrow) > 1.}

\item{new.flag}{Flag to plot the stamp in the current (if TRUE,
default) or new plotting window (if FALSE).}
}
\value{
Nothing
}
\description{
Add spict version to plot
}
