% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{validation.data.frame.old}
\alias{validation.data.frame.old}
\title{Collect results from the output of running validate.spict (this function is outdated).}
\usage{
validation.data.frame.old(ss)
}
\arguments{
\item{ss}{Output from validation.spict.}
}
\value{
A data frame containing the formatted validation results.
}
\description{
Collect results from the output of running validate.spict (this function is outdated).
}

