% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{validation.data.frame}
\alias{validation.data.frame}
\title{Collect results from the output of running validate.spict.}
\usage{
validation.data.frame(ss)
}
\arguments{
\item{ss}{Output from validation.spict.}
}
\value{
A data frame containing the formatted validation results.
}
\description{
Collect results from the output of running validate.spict.
}
