% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meyermillar.R
\name{write.bug.file}
\alias{write.bug.file}
\title{Write the BUGS code to a text file}
\usage{
write.bug.file(priors, fn = "tmp.bug")
}
\arguments{
\item{priors}{List of priors, typically coming from inp$meyermillar$priors.}

\item{fn}{Filename of to put BUGS code in.}
}
\value{
Nothing.
}
\description{
Write the BUGS code to a text file
}

