% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{acf.signf}
\alias{acf.signf}
\title{Check whether ACF of residuals is significant in any lags.}
\usage{
acf.signf(resid, lag.max = 4, return.p = FALSE)
}
\arguments{
\item{resid}{Vector of residuals.}

\item{lag.max}{Only check from lag 1 until lag.max.}

\item{return.p}{Return p-values of the calculated lags.}
}
\value{
Vector of TRUE and FALSE indicating whether significant lags were present. If return.p is TRUE then p-values are returned instead.
}
\description{
Check whether ACF of residuals is significant in any lags.
}
\details{
This corresponds to plotting the ACF using acf() and checking whether any lags has an acf value above the CI limit.
}
