% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.col.legend}
\alias{add.col.legend}
\title{Add a legend explaining colors of points (vertical orientation)}
\usage{
add.col.legend()
}
\value{
Nothing.
}
\description{
Add a legend explaining colors of points (vertical orientation)
}
