% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add.manlines}
\alias{add.manlines}
\title{Add lines to plot indicating result of management scenarios.}
\usage{
add.manlines(rep, par, par2 = NULL, index.shift = 0,
  plot.legend = TRUE, ...)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{par}{The name of the parameter to be plotted.}

\item{par2}{If a second parameter should be used as explanatory variable instead of time.}

\item{index.shift}{Shift initial time point by this index.}
}
\value{
Nothing
}
\description{
Add lines to plot indicating result of management scenarios.
}
