% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annual}
\alias{annual}
\title{Convert from quarterly (or other sub-annual) data to annual means, sums or a custom function.}
\usage{
annual(intime, vec, type = "mean")
}
\arguments{
\item{intime}{A time vector corresponding to the values in vec.}

\item{vec}{The vector of values to convert to annual means}

\item{type}{If type='mean' then annual mean is calculated, if type='sum' then annual sum is calculated. If type is a function, that function is used.}
}
\value{
A list containing the annual means and a corresponding time vector.
}
\description{
Convert from quarterly (or other sub-annual) data to annual means, sums or a custom function.
}
