% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{calc.influence}
\alias{calc.influence}
\title{Calculates influence statistics of observations.}
\usage{
calc.influence(rep, mc.cores = 1)
}
\arguments{
\item{rep}{A valid result from fit.spict().}

\item{mc.cores}{Number of cores for \code{parallel::mclapply} function. By
default 1.}
}
\value{
A list equal to the input with the added key "infl" containing
    influence statistics.
}
\description{
Calculates influence statistics of observations.
}
\details{
TBA
}
