% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{calc.influence}
\alias{calc.influence}
\title{Calculates influence statistics of observations.}
\usage{
calc.influence(rep)
}
\arguments{
\item{rep}{A valid result from fit.spict().}
}
\value{
A list equal to the input with the added key "infl" containing influence statistics.
}
\description{
Calculates influence statistics of observations.
}
\details{
TBA
}
