% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{calc.mase}
\alias{calc.mase}
\title{Calculate Mean Absolute Scaled Error (MASE)}
\usage{
calc.mase(rep, verbose = TRUE)
}
\arguments{
\item{rep}{Result of \code{\link{fit.spict}} that contains hindcasted runs added by \code{\link{hindcast}}.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
A data frame with estimate the MASE and the number of runs used for
    the estimation for each index.
}
\description{
Calculate Mean Absolute Scaled Error (MASE)
}
\details{
This function calculates the Mean Absolute Scaled Error (MASE) for
    each index time series by hindcasting. Thus, the application of this
    method requires a fitted spict object with the results of the hindcasting
    analysis \code{\link{hindcast}}.

    The smaller MASE, the higher the predictive power of the spict model
    regarding the prediction of index observations. In contrast, a MASE above
    1 suggests that the naive prediction of the index observations assuming
    the preceding index observations have a higher predictive power than the
    spict model. Note, however, that the absolute MASE value depends on
    multiple factors such as the number of peels, assumed priors, etc.

    Note that a difference in the timing of the index observations of less
    than a month are considered acceptable for the estimation of the naive
    prediction residuals and no warning is printed. If the variability
    exceeds a month, the predictions are still calculated, but a warning is
    printed.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep)
calc.mase(rep)
}
\references{
Carvalho, F., Winker, H., Courtney, D., Kapur, M., Kell, L., Cardinale, M.,
Schirripa, M., Kitakado, T., Yemane, D., Piner, K.R. Maunder, M.N., Taylor,
I., Wetzel, C.R., Doering, K., Johnsonm, K.F., Methot, R. D. (2021). A
cookbook for using model diagnostics in integrated stock assessments.
Fisheries Research, 240, 105959.

Kell, L. T., Kimoto, A., & Kitakado, T. (2016). Evaluation of the prediction
skill of stock assessment using hindcasting. Fisheries research, 183,
119-127.

Kell, L. T., Sharma, R., Kitakado, T., Winker, H., Mosqueira, I., Cardinale,
M., & Fu, D. (2021). Validation of stock assessment methods: is it me or my
model talking?. ICES Journal of Marine Science, 78(6), 2244-2255.

Winker, H., Carvalho, F., & Kapur, M. (2018). JABBA: just another Bayesian
biomass assessment. Fisheries Research, 204, 275-288.
}
