% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.om}
\alias{calc.om}
\title{Calculates the order of magnitude for the relative reference
    levels B/Bmsy and F/Fmsy}
\usage{
calc.om(rep)
}
\arguments{
\item{rep}{Result of fit.spict().}
}
\value{
Matrix containing the order of magnitude for B/Bmsy and
    F/Fmsy.
}
\description{
Calculates the order of magnitude for the relative reference
    levels B/Bmsy and F/Fmsy
}
\details{
The lower, upper values and the CI range are based on the
    95% confidence interval (CI).
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
