% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.process.resid}
\alias{calc.process.resid}
\title{Calculate process residuals}
\usage{
calc.process.resid(rep, dt = NULL)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{dt}{Time resolution of process residuals. By default (NULL), the
process residuals are calculated corresponding to the time resolution of
the input data.}
}
\value{
Data frame with year, process residuals for biomass and fishing
    mortality in the columns.
}
\description{
Calculate process residuals
}
\details{
Calculates process residuals for biomass and fishing mortality
    process.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- calc.process.resid(rep)
plotspict.diagnostic.process(rep)

}
