% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.process.resid}
\alias{calc.process.resid}
\title{Estimate process residuals}
\usage{
calc.process.resid(rep, dt = NULL)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{dt}{NULL by default}
}
\value{
Data frame with year, process residuals for biomass and fishing
    mortality.
}
\description{
Estimate process residuals
}
