% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{calc.tac}
\alias{calc.tac}
\title{Calculate Total Allowable Catch (TAC)}
\usage{
calc.tac(rep, inp = NULL, fractileCatch = 0.5, exp = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{inp}{Input list with ffac or catch observation corresponding to
management. If \code{NULL} (default), the input list in \code{rep} is
used.}

\item{fractileCatch}{The fractile of the catch distribution to be used for
setting the TAC. Default (0.5) corresponds to the median.}

\item{exp}{Should tac be reported on natural scale? (default: TRUE)}
}
\value{
Total Allowable Catch (TAC)
}
\description{
Calculate Total Allowable Catch (TAC)
}
