% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.tac}
\alias{calc.tac}
\title{Calculate TAC for set fishing mortality factor}
\usage{
calc.tac(rep, ffac, fracc = 0.5)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{ffac}{Fishing mortality factor}

\item{fracc}{The fractile of the catch distribution to be used
for setting the TAC. Default is median (0.5).}
}
\value{
Estimated TAC given fishing mortality factor
}
\description{
Calculate TAC for set fishing mortality factor
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
