% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{calc.tac}
\alias{calc.tac}
\title{Calculate Total Allowable Catch (TAC)}
\usage{
calc.tac(rep, ffac = NULL, fractileCatch = 0.5, exp = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{ffac}{Factor to multiply current F by}

\item{fractileCatch}{The fractile of the catch distribution to be used for
setting the TAC. Default (0.5) corresponds to the median.}

\item{exp}{Should tac be reported on natural scale? (default = TRUE)}
}
\value{
Total Allowable Catch (TAC)
}
\description{
Calculate Total Allowable Catch (TAC)
}
