% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{check.catchList}
\alias{check.catchList}
\title{Check catch list}
\usage{
check.catchList(catchList, sdfac = 1)
}
\arguments{
\item{catchList}{list obtaining the elements 'obsC', 'timeC', and 'dtc'
(optional element 'stdevfacC' which is 1 if not provided)}

\item{scenarioTitle}{title of scenario}
}
\value{
Checked catchList
}
\description{
Check catch list
}
