% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check.euler}
\alias{check.euler}
\title{Check robustness of assessment results vs time discretization}
\usage{
check.euler(rep, dteuler = 1/128)
}
\arguments{
\item{rep}{Result of fit.spict().}

\item{dteuler}{Vector with dteuler time steps to check against}
}
\value{
Relative difference between dteuler for reference levels and states
}
