% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{check.man.time}
\alias{check.man.time}
\title{Checks if management time is within model time}
\usage{
check.man.time(rep, verbose = TRUE)
}
\arguments{
\item{rep}{rep list}
}
\value{
rep
}
\description{
Checks if management time is within model time
}
