% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{check.man.time}
\alias{check.man.time}
\title{Checks if management time is within model time}
\usage{
check.man.time(rep, maninterval = NULL, maneval = NULL, verbose = TRUE)
}
\arguments{
\item{rep}{rep list}

\item{maninterval}{man interval}

\item{maneval}{man eval}

\item{verbose}{more output}
}
\value{
rep
}
\description{
Checks if management time is within model time
}
