% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.ffac}
\alias{get.ffac}
\title{Estimate fishing mortality factor minimising probability of specified
    model variable hitting a specified reference level under given fishing
    mortality}
\usage{
get.ffac(
  rep,
  var = "logBpBmsy",
  ref = 1,
  problevel = 0.95,
  reportmode = 1,
  verbose = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{var}{A variable of the spict model (default: "logBpBmsy").}

\item{ref}{Reference level relative to specified variable (default: 1)}

\item{problevel}{Probability level of the risk aversion (default: 0.95).}

\item{reportmode}{Integer between 0 and 2 determining which objects will be
adreported (default: 1).}

\item{verbose}{logical; print informative text (default: TRUE).}
}
\value{
Optimised Fishing mortality for P(Bp<Blim)
}
\description{
Estimate fishing mortality factor minimising probability of specified
    model variable hitting a specified reference level under given fishing
    mortality
}
