% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.ffac}
\alias{get.ffac}
\title{Get the fishing mortality corresponding to set probability
    of being above certain biomass threshold}
\usage{
get.ffac(rep, bfrac = 0.3, prob = 0.95, quant = "logBpBmsynotS",
  reportmode = 2)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{bfrac}{Fraction of biomass relativ to biomass reference
levels (dependent on \code{quant}), e.g.  fraction of B/Bmsy
which is defined as threshold (Blim = 0.3 Bmsy, Btrigger = 0.5
Bmsy) or fraction of Bp/Bl}

\item{prob}{Probability to be above threshold (bfrac; default
is 0.95)}

\item{quant}{Quantity to use (options: "logBpBmsy", logBpBl")}

\item{reportmode}{default = 2}
}
\value{
Optimised Fishing mortality for P(Bp<Blim)
}
\description{
Get the fishing mortality corresponding to set probability
    of being above certain biomass threshold
}
