% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.TAC}
\alias{get.TAC}
\title{Estimate the Total Allowable Catch (TAC)}
\usage{
get.TAC(rep, hcr = "MSY", args = NULL, curc = NULL, sdfac = 0.001,
  getFit = FALSE)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{hcr}{SPiCT specific harvest control rule (HCR). Possible
rules are: \code{"MSY"}, \code{"MSY-HS"}, or \code{"MSY-HS-PA"}
(see details for more information).  below.}

\item{args}{A list with specific arguments for the respective HCR
(see details for possible arguments).}

\item{curc}{Optional; catch during assessment year, e.g. last
year's TAC (default: \code{NULL}; see details for more
information).}

\item{sdfac}{Factor for the multiplication of the standard
deveiation of the catch during the assessment year
(\code{stdevfacC}; default = 1e-3).}

\item{getFit}{Logical; if \code{TRUE} the function returns the
fitted 'spictcls' object with respective HCR (\code{FALSE} by
default).}
}
\value{
A list with absolute and relative reference levels and
    states and estimated TAC; if \code{getFit} is \code{TRUE} the
    fitted object with the respective HCR is returned.
}
\description{
Estimate the Total Allowable Catch (TAC)
}
\examples{
rep <- fit.spict(pol$albacore)
get.TAC(rep)
}
\references{
ICES. 2017. Report of the Workshop on the Development of the ICES
approach to providing MSY advice for category 3 and 4 stocks
(WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/
ACOM:47. 53 pp.
 
ICES. 2018. Report of the Eighth Workshop on the Development of
Quantitative Assessment Methodologies based on LIFE-history traits,
exploitation characteristics, and other relevant parameters for
data-limited stocks (WKLIFE VIII), 8-12 October 2018, Lisbon,
Portugal. ICES CM 2018/ACOM:40. 172 pp.

ICES 2019. Report of the Ninth Workshop on the Development of
Quantitative Assessment Methodologies based on LIFE-history traits,
exploitation characteristics, and other relevant parameters for
data-limited stocks (WKLIFE IX), 30 September-4 October 2019,
Lisbon, Portugal.
}
