% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.TAC}
\alias{get.TAC}
\title{Calculate TAC based on specific HCR}
\usage{
get.TAC(
  rep,
  scenarioTitle = "",
  maninterval = NULL,
  maneval = NULL,
  ffac = NULL,
  cfac = NULL,
  csdfac = 1,
  fractiles = list(catch = 0.5, bbmsy = 0.5, ffmsy = 0.5),
  breakpointB = 0,
  safeguardB = list(limitB = 0, prob = 0.95),
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  catchList = NULL,
  verbose = TRUE,
  dbg = 0,
  mancheck = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{scenarioTitle}{Title of scenario (default: \code{'customScenario_X'},
where X corresponds to a numeric counting equal to the numer of scenarios
with this name in \code{rep$man} plus 1, e.g. \code{'customScenario_3'}).}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: \code{maninterval = c(2020.25,2021.25)}.
Default: NULL.}

\item{maneval}{Time at which to evaluate model states. Example: \code{maneval
= 2021.25}. Default: NULL.}

\item{ffac}{Factor to multiply current fishing mortality by (default: 1).}

\item{cfac}{Factor to multiply current catch by (default: 1). Please refer to
the details for more information.}

\item{csdfac}{Factor for the multiplication of the standard deviation of the
catch (default: 1). Please refer to the details for more information.}

\item{fractiles}{List defining the fractiles of the 3 distributions of
'catch', 'bbmsy', and 'ffmsy'. By default (0.5) median is used for all 3
quantities. Please refer to the details for more information.}

\item{safeguardB}{List defining an optional precautionary buffer by means of
a biomass reference level relative to \eqn{B/B_{MSY}} (\code{'limitB'};
default: 0, i.e. deactivating the PA buffer) and the risk aversion
probability (\code{'prob'}; default: 0.95). Please refer to the details
for more information.}

\item{intermediatePeriodCatch}{Catch during intermediate period, e.g. last
year's TAC (default: \code{NULL}). Please refer to the details for more
information.}

\item{intermediatePeriodCatchSDFac}{Factor for the multiplication of the
standard deviation of the catch during the intermediate period (default:
1). Please refer to the details for more information.}

\item{intermediatePeriodCatchList}{List defining catch in the intermediate
period obtaining the elements 'obsC', 'timeC', and 'dtc' (optional
element 'stdevfacC' which is 1 if not provided). Please refer to the
details for more information.}

\item{catchList}{List obtaining the elements 'obsC', 'timeC', and 'dtc'
(optional element 'stdevfacC' which is 1 if not provided). Please refer
to the details for more information.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}

\item{mancheck}{Should the time-dependent objects in \code{inp} be checked
against the management time and corrected if necessary? (Default: TRUE)}

\item{breakpointBBmsy}{Breakpoint in terms of \eqn{B/B_{MSY}} for the
hockey-stick HCR. By default (0) no breakpoint is assumed. Please refer
to the details for more information.}
}
\value{
The total allowable catch based on specified HCR.
}
\description{
Calculate TAC based on specific HCR
}
\details{
For detailed information, please refer to
    \code{\link{add.man.scenario}}.
}
\examples{
rep <- fit.spict(pol$albacore)

## Fishing at Fmsy
get.TAC(rep)

## MSY hockey-stick rule
get.TAC(rep, breakpointB = 0.5)

## ICES (2019) recommended HCR
get.TAC(rep, fractiles = list(catch=0.35, bbmsy=0.35, ffmsy=0.35), breakpointB=0.5)

}
