% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.TAC}
\alias{get.TAC}
\title{Estimate the Total Allowable Catch (TAC)}
\usage{
get.TAC(rep, hcr = "MSY", args = NULL, getFit = FALSE)
}
\arguments{
\item{rep}{Result list from fit.spict().}

\item{hcr}{SPiCT specific harvest control rule (HCR). Possible
rules are: \code{"MSY"}, \code{"MSY-PA"}, \code{"Btrend"}, or
\code{"MSY-Btrend"}. More details about the rules are provided
below.}

\item{args}{A list with specific arguments for the respective
HCR. More details are provided below.}

\item{getFit}{Logical; if \code{TRUE} the function returns the
fitted 'spictcls' object with respective HCR (\code{FALSE} by
default).}
}
\value{
A list with absolute and relative reference levels and
    states and estimated TAC; if \code{getFit} is \code{TRUE} the
    fitted object with the respective HCR is returned.
}
\description{
Estimate the Total Allowable Catch (TAC)
}
\details{
The possible harvest control rules (HCRs) are:
\itemize{
  \item{"MSY"- ICES MSY approach.}
  \item{"MSY-PA" - MSY precautionary approach.}
  \item{"Btrend" - Data-limited approach based on the trend in the biomass.}
  \item{"MSY-Btrend" - Combined approach of the MSY and the biomass
trend approach. Switching between the two based on the order of magnitude of
the confidence intervals of F/F_{MSY} and B/B_{MSY}.}
}

The arguments of the 'args' list are:
\itemize{
  \item{fracc - Fractile of the predicted catch distribution (default: 0.5)}
  \item{fracf - Fractile of the \eqn{F/F_{MSY}} distribution (default: 0.5)}
  \item{fracb - Fractile of the \eqn{B/B_{MSY}} distribution (default: 0.5)}
  \item{bfrac - }
  \item{babs - }
  \item{om - }
  \item{btrend - }
  \item{reportmode - }
}
}
\examples{
rep <- fit.spict(pol$albacore)
get.TAC(rep)
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
