% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get.TAC}
\alias{get.TAC}
\title{Estimate Total Allowable Catch (TAC)}
\usage{
get.TAC(rep, hcr = "MSY", args = NULL, getFit = FALSE)
}
\arguments{
\item{rep}{Result list as output from fit.spict().}

\item{hcr}{Harvest control rule. Options: 'msy', 'pa', 'dl', '2/3'
(more information under details)}

\item{args}{A list with arguments for the individual hcrs}

\item{getFit}{Logical; if TRUE the fitted results list with
adjusted fsihing mortality value is returned. Default is FALSE.}
}
\value{
A list with estimated TAC based on harvest control rule
    settings or the fitted rep list with adjusted fishing mortality
    values if getFit = TRUE and a logical value indicating if the
    stability clause was hit or not (if in use).
}
\description{
Estimate Total Allowable Catch (TAC)
}
\details{
The possible harvest control rules are:
\itemize{
  \item{"MSY"}{ICES MSY approach}
  \item{"MSY-PA"}{MSY precautionary approach}
  \item{"Btrend"}{Data-limited approach based on the trend in the biomass}
  \item{"MSY-Btrend"}{Data-limited approach based on the trend in the biomass}
}
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
get.TAC(rep)
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
