% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.TACi}
\alias{get.TACi}
\title{Apply uncertainty cap to F or F multiplier}
\usage{
get.TACi(repin, ffac, fractileC = 0.5)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{ffac}{Absolute F value for which to predict the catch for}

\item{fractileC}{fractileC The fractile of the catch distribution to be used for setting TAC. Default
is median (0.5).}
}
