% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{hindcast}
\alias{hindcast}
\title{Conduct hindcasting analysis}
\usage{
hindcast(
  rep,
  npeels = 7,
  reduce.output.size = TRUE,
  mc.cores = 1,
  peel.dtc = FALSE
)
}
\arguments{
\item{rep}{rep Result of \code{\link{fit.spict}}.}

\item{npeels}{Number of years/seasons (dependent on dtc) of data (catch and
effort) to remove (this is also the total number of model runs).}

\item{reduce.output.size}{logical, if \code{TRUE} (default) hindcasting is
run with \code{getReportCovariance} and \code{getJointPrecision} set as
\code{FALSE}}

\item{mc.cores}{Number of cores for \code{parallel::mclapply} function. By
default 1.}

\item{peel.dtc}{Peel according to catch seasons (dtc) rather than years? It
only differs if the data includes seasonal catches (Default: FALSE)}
}
\value{
A \code{spictcls} list with the added element \code{hindcast}
    containing the results of the hindcasting analysis. Use
    \code{\link{plotspict.hindcast}} to plot these results.
}
\description{
Conduct hindcasting analysis
}
\details{
This method creates a number of subsets (or peels) specified with
    the argument \code{npeels} by sequentially omitting all observations from
    the most recent time step (by default corresponding to a year). Then,
    spict is fitted to each subset while excluding all index observations in
    the most recent year with data. The resulting fitted spict objects are
    attached to the base spict object as a list element labeled
    \code{hindcast} and resulted by this method.

    The prediction of the excluded index observations relative to the actual
    excluded index can then be compared to a 'naive' prediction using the
    preceding index observation. This allows to estimate the Mean Absolute
    Scaled Error (MASE) for each index \code{\link{calc.mase}}.

    The predicted indices can be visualised with the function
    \code{\link{plotspict.hindcast}}.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep, npeels = 5)
plotspict.hindcast(rep)
}
\references{
Carvalho, F., Winker, H., Courtney, D., Kapur, M., Kell, L., Cardinale, M.,
Schirripa, M., Kitakado, T., Yemane, D., Piner, K.R. Maunder, M.N., Taylor,
I., Wetzel, C.R., Doering, K., Johnsonm, K.F., Methot, R. D. (2021). A
cookbook for using model diagnostics in integrated stock assessments.
Fisheries Research, 240, 105959.

Kell, L. T., Kimoto, A., & Kitakado, T. (2016). Evaluation of the prediction
skill of stock assessment using hindcasting. Fisheries research, 183,
119-127.

Kell, L. T., Sharma, R., Kitakado, T., Winker, H., Mosqueira, I., Cardinale,
M., & Fu, D. (2021). Validation of stock assessment methods: is it me or my
model talking?. ICES Journal of Marine Science, 78(6), 2244-2255.

Winker, H., Carvalho, F., & Kapur, M. (2018). JABBA: just another Bayesian
biomass assessment. Fisheries Research, 204, 275-288.
}
