% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spict.R
\name{make.obj}
\alias{make.obj}
\title{Create TMB obj using TMB::MakeADFun and squelch screen printing.}
\usage{
make.obj(datin, pl, inp, phase = 1)
}
\arguments{
\item{datin}{Data list.}

\item{pl}{Parameter list.}

\item{inp}{List of input variables as output by check.inp.}

\item{phase}{Estimation phase, integer.}
}
\value{
List to be used as data input to TMB.
}
\description{
Create TMB obj using TMB::MakeADFun and squelch screen printing.
}
