% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.select}
\alias{man.select}
\title{Select management scenarios}
\usage{
man.select(rep, scenarios = "all", spictcls = FALSE, verbose = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or
\code{add.man.scenario}.}

\item{scenarios}{Selection of scenarios in preferred order. Can be a vector
with the names of the selected scenarios or numbers indicating their
position in \code{rep$man}, e.g. \code{c(6,2)} for the 6th and 2nd
scenario in the \code{rep$man} list. Setting this argument to \code{NULL}
or "none", removes all scenarios from the spict object. By default
(\code{'all'}), all scenarios are selected.}

\item{spictcls}{Should selected scenario be a standard spictcls object?
Default is \code{FALSE}. See details for more information.}

\item{verbose}{Should detailed outputs be provided (default: \code{TRUE}).}
}
\value{
A fitted spict object wit selected management scenarios in preferred
    order in \code{rep$man}. This function can also be used to select a
    specific scenarios in \code{rep$man} as the new main spictcls object. By
    setting the argument \code{spictcls} to \code{TRUE}, management related
    catch observations are removed and the retaped spict object of class
    'spictcls' is returned, comparable to the object returned by
    \code{\link{fit.spict}}. This only works if one scenario is selected
    (\code{length(scnearios) == 1}).
}
\description{
Select management scenarios
}
\examples{

data(pol)
rep <- fit.spict(pol$albacore)
rep <- manage(rep, c(2,4,6))

## based on names
names(rep$man)
rep1 <- man.select(rep, c("currentF","noF"))

## based on indices
length(rep$man)
rep2 <- man.select(rep, c(1,3))

## select specific scenario as new spictcls object
rep3 <- man.select(rep, 1, spictcls = TRUE)

}
