% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.select}
\alias{man.select}
\title{Select management scenarios}
\usage{
man.select(rep, scenarios = "all")
}
\arguments{
\item{rep}{Result list from manage().}

\item{scenarios}{scenarios to select in preferred order}
}
\value{
rep wit selected management scenarios
}
\description{
Select management scenarios
}
