% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.tac}
\alias{man.tac}
\title{Get the TAC for the management scenarios}
\usage{
man.tac(rep, fractileCatch = 0.5, exp = TRUE)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or \code{man.scenario}.}

\item{fractileCatch}{fractile}

\item{exp}{Should tac be reported on natural scale? (default = TRUE)}
}
\value{
rep wit selected management scenarios
}
\description{
Get the TAC for the management scenarios
}
