% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{man.tac}
\alias{man.tac}
\title{Get the TAC for the management scenarios}
\usage{
man.tac(rep)
}
\arguments{
\item{rep}{Result list from manage().}
}
\value{
rep wit selected management scenarios
}
\description{
Get the TAC for the management scenarios
}
