% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{manage}
\alias{manage}
\title{Calculate predictions under 8 default management scenarios}
\usage{
manage(
  rep,
  scenarios = "all",
  maninterval = NULL,
  maneval = NULL,
  intermediatePeriodCatch = NULL,
  intermediatePeriodCatchSDFac = 1,
  intermediatePeriodCatchList = NULL,
  verbose = TRUE,
  dbg = 0
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{scenarios}{Vector of integers specifying which scenarios to run or
'all' to run all scenarios. Default: 'all'.}

\item{maninterval}{Two floats representing the start and end of the
management period. Example: maninterval = c(2020.25,2021.25). Default:
NULL.}

\item{maneval}{Time at which to evaluate model states. Example: maneval =
2021.25. Default: NULL.}

\item{intermediatePeriodCatch}{Catch during intermediate period, e.g. last year's TAC (default:
\code{NULL}; see details for more information).}

\item{intermediatePeriodCatchSDFac}{Factor for the multiplication of the standard deviation of the
catch during the intermediate period (default: 1).}

\item{intermediatePeriodCatchList}{List defining catch in the intermediate
period obtaining the elements 'obsC', 'timeC', and 'dtc' (optional
element 'stdevfacC' which is 1 if not provided)}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}
}
\value{
List containing results of management calculations.
}
\description{
Calculate predictions under 8 default management scenarios
}
\details{
The 8 default scenarios are:

\itemize{
  \item{"1"}{"currentCatch": Keep the catch of the current year (i.e. the last observed catch).}
  \item{"2"}{"currentF":  Keep the F of the current year.}
  \item{"3"}{"Fmsy": Fish at Fmsy i.e. F=Fmsy.}
  \item{"4"}{"noF": No fishing, reduce to 1\% of current F.}
  \item{"5"}{"reduceF25": Reduce F by X\%. Default X = 25.}
  \item{"6"}{"increaseF25": Increase F by X\%. Default X = 25.}
  \item{"7"}{"msyHockeyStick": Use ICES MSY hockey-stick advice rule (ICES, 2017).}
  \item{"8"}{"ices": Use ICES MSY 35th hockey-stick advice rule (ICES, 2019).}}

Scenario 7 implements the ICES MSY advice rule for stocks that are assessed
using spict (ICES 2017). MSY B_{trigger} is set equal to B_{MSY} / 2. Then
fishing mortality in the short forecast is calculated as:

F(y+1) = F(y) * min{ 1, median[B(y+1) / MSY B_{trigger}] } /
median[F(y)/F_{MSY}

Scenario 8 is similar to scenario 7, but includes assessment uncertainty in
the predictions by using the 35th percentile of the distributions of the
predicted catch, B/B_{MSY} and F/F_{MSY}.

Dependent on the start of the management period (e.g. advice year), there
might be a time lag between the last observation and the start of the
management period, often referred to as the intermediate period. If this is
the case, an assumption about the catch during intermediate time period (e.g.
assessment year) has to be made. Two meaningful assumptions are:

\itemize{
\item{1:}{The catch in the intermediate period is based on the fishing
mortality which is extrapolated from the previous year. This is the default
assumption;}
\item{2:}{The catch in the intermediate period is directly specified. This
could for example be the TAC recommended in the previous year. The catch can
be specified by means of the argument \code{intermediatePeriodCatch}. Be
aware that this catch might correspond to several years or a fraction of a
year depending on the time between the last observation and the start of the
management period. The function \code{\link{man.timeline}} can help
visualising the default or specified intermediate period in your data. The
argument \code{intermediatePeriodCatchSDFac} allows to specify the factor
with which to multiply the standard deviation of the catch (\eqn{\sigma_C})
with. It is thus a measure of the certainty around the catch in the
intermediate period. The argument \code{intermediatePeriodCatchList} allows
to define a list with catches and their intervals. It is a list with the
elements 'obsC', 'timeC', 'dtc' and the optional element 'stdevfacC' (which
is equal to 1 if not provided).}}
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
repman <- manage(rep, c(2,4,8))
sumspict.manage(repman) # To print projections

}
\references{
ICES. 2017. Report of the Workshop on the Development of the ICES
    approach to providing MSY advice for category 3 and 4 stocks
    (WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/ACOM:47.
    53 pp.
}
