% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{manage}
\alias{manage}
\title{Calculate predictions under different management scenarios}
\usage{
manage(
  repin,
  scenarios = "all",
  manstart = NULL,
  dbg = 0,
  catch = NULL,
  sdfac = 1,
  catchList = NULL
)
}
\arguments{
\item{repin}{Result list from fit.spict().}

\item{scenarios}{Vector of integers specifying which scenarios to run.
Default: 'all'.}

\item{manstart}{Year that management should be initiated.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}

\item{catch}{Catch which will be taken each year in the prediction period. By
default (\code{NULL}) catch is equal to last year's catch.}

\item{sdfac}{Take catch with this 'stdevfacC' (default = 1e-3)}

\item{catchList}{List with elements: "obsC", "timeC", "stdevfacC", and "dtc"
specifying the catch, the time of catch, the factor of the standard
deviation of the catch, and the catch time interval. By default
(\code{NULL}) settings are based on the last year.}
}
\value{
List containing results of management calculations.
}
\description{
Calculate predictions under different management scenarios
}
\details{
Scenarios that are currently implemented include:
\itemize{
  \item{"1"}{ Keep the catch of the current year (i.e. the last observed catch).}
  \item{"2"}{ Keep the F of the current year.}
  \item{"3"}{ Fish at Fmsy i.e. F=Fmsy.}
  \item{"4"}{ No fishing, reduce to 1\% of current F.}
  \item{"5"}{ Reduce F by X\%. Default X = 25.}
  \item{"6"}{ Increase F by X\%. Default X = 25.}
  \item{"7"}{ Use ICES MSY advice rule (ICES, 2017).}

Scenario 7 implements the ICES MSY advice rule for stocks that are assessed using spict (ICES 2017). MSY B_{trigger} is set equal to B_{MSY} / 2. Then fishing mortality in the short forecast is calculated as:

F(y+1) = F(y) * min{ 1, median[B(y+1) / MSY B_{trigger}] } /
median[F(y)/F_{MSY}] }
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
repman <- manage(rep)
mansummary(repman) # To print projections
}
\references{
{ICES. 2017. Report of the Workshop on the Development of the ICES approach to providing MSY advice for category 3 and 4 stocks (WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/ ACOM:47. 53 pp.}

ICES. 2017. Report of the Workshop on the Development of the ICES
    approach to providing MSY advice for category 3 and 4 stocks
    (WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/ACOM:47.
    53 pp.
}
