% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{manage}
\alias{manage}
\title{Calculate predictions under 8 default management scenarios}
\usage{
manage(
  rep,
  scenarios = "all",
  maninterval = NULL,
  maneval = NULL,
  verbose = TRUE,
  dbg = 0
)
}
\arguments{
\item{rep}{A result report as generated by running \code{fit.spict}.}

\item{scenarios}{Vector of integers specifying which scenarios to run or
'all' to run all scenarios. Default: 'all'.}

\item{maninterval}{Year that management should be initiated.}

\item{maneval}{when to evaluate management}

\item{verbose}{more output?}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more output.}
}
\value{
List containing results of management calculations.
}
\description{
Calculate predictions under 8 default management scenarios
}
\details{
The 8 default scenarios are:

\itemize{
  \item{"1"}{"currentCatch": Keep the catch of the current year (i.e. the last observed catch).}
  \item{"2"}{"currentF":  Keep the F of the current year.}
  \item{"3"}{"Fmsy": Fish at Fmsy i.e. F=Fmsy.}
  \item{"4"}{"noF": No fishing, reduce to 1\% of current F.}
  \item{"5"}{"redF25": Reduce F by X\%. Default X = 25.}
  \item{"6"}{"incrF25": Increase F by X\%. Default X = 25.}
  \item{"7"}{"msyHockeyStick": Use ICES MSY hockey-stick advice rule (ICES, 2017).}
  \item{"8"}{"ices": Use ICES MSY 35th hockey-stick advice rule (ICES, 2019).}}

Scenario 7 implements the ICES MSY advice rule for stocks that are assessed
using spict (ICES 2017). MSY B_{trigger} is set equal to B_{MSY} / 2. Then
fishing mortality in the short forecast is calculated as:

F(y+1) = F(y) * min{ 1, median[B(y+1) / MSY B_{trigger}] } /
median[F(y)/F_{MSY}

Scenario 8 is similar to scenario 7, but includes assessment uncertainty in
the predictions by using the 35th percentile of the distributions of the
predicted catch, B/B_{MSY} and F/F_{MSY}.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
repman <- manage(rep)
sumspict.manage(repman) # To print projections

}
\references{
ICES. 2017. Report of the Workshop on the Development of the ICES
    approach to providing MSY advice for category 3 and 4 stocks
    (WKMSYCat34), 6-10 March 2017, Copenhagen, Denmark. ICES CM 2017/ACOM:47.
    53 pp.
}
