% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{mansummary}
\alias{mansummary}
\title{Print management summary.}
\usage{
mansummary(rep, ypred = 1, include.EBinf = FALSE, verbose = TRUE)
}
\arguments{
\item{rep}{Result list as output from manage().}

\item{ypred}{Show results for ypred years into the future.}

\item{include.EBinf}{Include EBinf/Bmsy in the output.}

\item{verbose}{Print more details on observed and predicted time intervals.}
}
\value{
Data frame containing management summary.
}

