% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{mansummary}
\alias{mansummary}
\title{Print management summary.}
\usage{
mansummary(
  repin,
  ypred = 1,
  include.EBinf = FALSE,
  include.unc = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{repin}{Result list as output from manage().}

\item{ypred}{Show results for ypred years from manstart.}

\item{include.EBinf}{Include EBinf/Bmsy in the output.}

\item{include.unc}{Include uncertainty of management quantities.}

\item{verbose}{Print more details on observed and predicted time intervals.}
}
\value{
Data frame containing management summary.
}
\description{
Print management summary.
}
