% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.col}
\alias{plot.col}
\title{Plot model points colored depending on the quarter to which they belong.}
\usage{
\method{plot}{col}(time, obs, obsx = NULL, pch = 1, add = FALSE,
  typ = "p", do.line = TRUE, add.legend = FALSE,
  add.vline.at = NULL, ...)
}
\arguments{
\item{time}{Time vector.}

\item{obs}{Observation vector (or residual vector).}

\item{obsx}{Second observation vector for use as independent variable instead of time.}

\item{pch}{Point character.}

\item{add}{If TRUE plot is added to the current plot.}

\item{typ}{Plot type.}

\item{do.line}{If TRUE draw a line between points.}

\item{add.legend}{If TRUE add legend containing information on quarters.}

\item{add.vline.at}{If not NULL will draw a vertical line at the given time point.}

\item{...}{Additional plotting arguments.}
}
\value{
Nothing.
}
\description{
Plot model points colored depending on the quarter to which they belong.
}
