% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.spictcls}
\alias{plot.spictcls}
\title{3x3 plot illustrating spict results.}
\usage{
\method{plot}{spictcls}(rep, logax = FALSE)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of relevant axes? default: FALSE}
}
\value{
Nothing.
}
\description{
3x3 plot illustrating spict results.
}
\details{
Create a 3x3 plot containing the following:
\itemize{
 \item{1. Biomass using plotspict.biomass().}
 \item{2. One-step-ahead residuals, only if calculated, using plotspict.osar().}
 \item{3. One-step-ahead auto-correlation function (only if calculated).}
 \item{4. Estimated F versus estimated B using plotspict.fb().}
 \item{5. Estimated fishing mortality using plotspict.f().}
 \item{6. Observed versus predicted catches using plotspict.catch().}
 \item{7. Observed versus theoretical production using plotspict.production().}
 \item{8. Calculated time-constant using plotspict.tc().}
}
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plot(rep)
}

