% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.spictcls}
\alias{plot.spictcls}
\title{Plot summarising spict results.}
\usage{
\method{plot}{spictcls}(
  x,
  stamp = get.version(),
  verbose = TRUE,
  CI = 0.95,
  add.loess = FALSE,
  span = 0.75,
  ...
)
}
\arguments{
\item{x}{A result report as generated by running fit.spict.}

\item{stamp}{Stamp plot with this character string.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90%
confidence intervals. By default (CI = 0.95), the 95% confidence
intervals are estimated.}

\item{add.loess}{Add smooth line (using \code{loess}) to residuals (Default: FALSE).}

\item{span}{Parameter that controls the degree of smoothing (only used if \code{add.loess=TRUE}, Default: 0.75).}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Nothing.
}
\description{
Plot summarising spict results.
}
\details{
Create a plot containing the following:
\itemize{
 \item{1. Estimated biomass using plotspict.biomass().}
 \item{2. Estimated fishing mortality using plotspict.f().}
 \item{3. Observed versus predicted catches using plotspict.catch().}
 \item{4. Estimated biomass relative to Bmsy using plotspict.bbmsy().}
 \item{5. Estimated fishing mortality relative to Fmsy using plotspict.ffmsy().}
 \item{6. Estimated F versus estimated B using plotspict.fb().}
 \item{7. Observed versus theoretical production using plotspict.production().}
}

Optional plots included if relevant:
\itemize{
 \item{ Estimated seasonal spline using plotspict.season().}
 \item{ Calculated time-constant using plotspict.tc().}
 \item{ First prior and corresponding posterior distribution using plotspict.priors().}
 \item{ One-step-ahead residuals of catches using plotspict.osar().}
 \item{ One-step-ahead residuals of catches using plotspict.osar().}
}

If no management scenarios are included in \code{rep$man}, the grey vertical
line corresponds to the time of the last observation. If management scenarios
are included in \code{rep$man}, the prediction and confidence intervals of
the base scenario (\code{rep}) are omitted and instead the projections of the
different management scenarios are drawn in different colours. Dotted lines
of the management scenarios reflect the intermediate period, while solid
lines reflect the management period. Additionally, two vertical lines
correspond to the start and end of the management period.

Be aware that potential catch intervals of more a year, e.g. biennial
assessment so that the intermediate period spans two years, or management
period spans two years, are equally split up into annual intervals.

Be aware of the fact that the catches represent intervals, where the length
of the interval is indicated by \code{dtc}, e.g. with \eqn{dtc = 1, C(1990) =
[1990,1990[}. In the plot the catches (and vertical lines) correspond to the
beginning of the catch interval. It might thus seem as if the time of the
vertical lines and the management interval would not align.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plot(rep)

}
