% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.bbmsy}
\alias{plotspict.bbmsy}
\title{Plot estimated B/Bmsy.}
\usage{
plotspict.bbmsy(rep, logax = FALSE, main = "Relative biomass",
  ylim = NULL, plot.obs = TRUE, qlegend = TRUE, lineat = 1,
  xlab = "Time", stamp = get.version())
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}

\item{main}{Title of plot.}

\item{ylim}{Limits for y-axis.}

\item{plot.obs}{If TRUE observations are plotted.}

\item{qlegend}{If TRUE legend explaining colours of observation data is plotted.}

\item{lineat}{Draw horizontal line at this y-value.}

\item{xlab}{Label of x-axis.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\description{
Plot estimated B/Bmsy.
}
\details{
Plots estimated B/Bmsy.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.bbmsy(rep)
}
