% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.btrend}
\alias{plotspict.btrend}
\title{Plot the expected biomass trend}
\usage{
plotspict.btrend(rep, CI = 0.95)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}
}
\value{
Nothing.
}
\description{
Plot the expected biomass trend
}
