% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.ci}
\alias{plotspict.ci}
\title{Plot catch and index data.}
\usage{
plotspict.ci(inp, stamp = get.version())
}
\arguments{
\item{inp}{An input list containing data.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing
}
\description{
Plot catch and index data.
}
