% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.compare}
\alias{plotspict.compare}
\title{Compare different spict fits}
\usage{
plotspict.compare(
  rep,
  ...,
  varname = c("B", "F", "C", "BBmsy", "FFmsy", "P"),
  exp = TRUE,
  CI = 0.95,
  plot.unc = TRUE,
  col = c("dodgerblue2", "darkorange1", "forestgreen", "goldenrod1", "purple2",
    "firebrick3", "skyblue4", "darkgreen", "salmon3", "brown2"),
  asp = 2,
  plot.legend = TRUE,
  stamp = get.version()
)
}
\arguments{
\item{rep}{A result report as generated by running \code{\link{fit.spict}} or
a list containing objects fitted with spict.}

\item{...}{Optional additional spict fits.}

\item{varname}{Name of the variable to be plotted. The following options are
currently implemented: \code{"B"} for the biomass, \code{"F"} for the
fishing mortality, \code{"C"} for the catch, \code{"BBmsy"} for the
biomass, \code{"FFmsy"} for the fishing mortality, and \code{"P"} for the
surplus production.}

\item{exp}{Logical; indicating whether to plot the results on the natural or
logarithmic scale. By default (\code{TRUE}), results are plotted in the
natural scale.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90%
confidence intervals. By default (\code{0.95}), the 95% confidence
intervals are estimated and plotted.}

\item{plot.unc}{Logical or integer indicating whether to include the
uncertainty intervals (CIs). If equal to \code{TRUE} (or \code{1}), CIs
are drawn as dashed lines, if equal to \code{2}, CIs are drawn as shaded
polygons. Default: \code{TRUE}.}

\item{col}{Colours for different spict fits.}

\item{asp}{Positive number deining the target aspect ratio (columns / rows)
of the plot arrangement.}

\item{plot.legend}{Logical; Indicating whether to include a legend. Default:
\code{TRUE}. Use \code{plot.legend = 2} to plot the legend inside the
last panel.}

\item{stamp}{Stamp plot with this character string.}
}
\description{
Compare different spict fits
}
\details{
This function plots the results of spict fits in a single plot
}
\examples{
data(pol)
inp <- pol$albacore
inp$dteuler <- 1/8
rep1 <- fit.spict(inp)
inp$priors$logbkfrac <- c(log(0.3), 0.1, 1)
rep2 <- fit.spict(inp)
plotspict.compare(list(def = rep1, bkprior = rep2))

}
