% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.diagnostic.process}
\alias{plotspict.diagnostic.process}
\title{Plot model diagnostics regarding processes and process residuals}
\usage{
plotspict.diagnostic.process(
  rep,
  lag.max = 4,
  qlegend = TRUE,
  plot.data = TRUE,
  mfcol = FALSE,
  add.loess = FALSE,
  span = 0.75,
  stamp = get.version()
)
}
\arguments{
\item{rep}{A result report as generated by running \code{\link{fit.spict}}
that contains process residuals calculated by the function
\code{\link{process.resid}}.}

\item{lag.max}{Maximum lag to use in acf calculations.}

\item{qlegend}{If TRUE plot a legend showing quarter of year information.}

\item{plot.data}{If TRUE plot data in the top row (this option is only applied if osa residuals have been calculated).}

\item{mfcol}{If TRUE plot plots columnwise (FALSE => rowwise).}

\item{add.loess}{Add smooth line (using \code{loess}) to residuals (Default: FALSE).}

\item{span}{Parameter that controls the degree of smoothing (only used if \code{add.loess=TRUE}, Default: 0.75).}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Invisible \code{NULL}
}
\description{
Plot model diagnostics regarding processes and process residuals
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- calc.process.resid(rep)
plotspict.diagnostic.process(rep)


}
\seealso{
\code{\link{process.resid}}
}
