% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.diagnostic}
\alias{plotspict.diagnostic}
\title{Plot model diagnostic (data, residuals, and more)}
\usage{
plotspict.diagnostic(rep)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}
}
\value{
Nothing.
}
\description{
Plot model diagnostic (data, residuals, and more)
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
rep <- calc.osa.resid(rep)
plotspict.diagnostic(rep)
}

