% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.growth}
\alias{plotspict.growth}
\title{Plot estimated time-varying growth}
\usage{
plotspict.growth(rep, logax = FALSE, main = "Time-varying growth",
  ylim = NULL, xlim = NULL, xlab = "Time", plot.ci = TRUE,
  stamp = get.version())
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{logax}{Take log of y-axis? default: FALSE}

\item{main}{Title of plot.}

\item{ylim}{Limits for y-axis.}

\item{xlim}{Limits for x-axis.}

\item{xlab}{Label of x-axis.}

\item{plot.ci}{If TRUE 95\% CIs are included.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\details{
Plots estimated time-varying growth
}

