% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.hcr}
\alias{plotspict.hcr}
\title{Plot harvest control rules}
\usage{
plotspict.hcr(rep, stamp = get.version())
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\description{
Plot harvest control rules
}
\details{
Plots harvest control rules
}
