% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.hindcast}
\alias{plotspict.hindcast}
\title{Hindcasting plot for indices}
\usage{
plotspict.hindcast(
  rep,
  add.mase = TRUE,
  CI = 0.95,
  verbose = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = "Year",
  ylab = NA,
  plot.log = TRUE,
  mfrow = NULL,
  mar = c(2, 2, 3, 1) + 0.1,
  oma = c(3, 3, 1, 1),
  legend.title = NULL,
  legend.pos = "topright",
  legend.ncol = 1,
  asp = 2,
  stamp = get.version()
)
}
\arguments{
\item{rep}{A result report as generated by running \code{\link{fit.spict}}
that contains hindcasted runs added by \code{\link{hindcast}}.}

\item{add.mase}{Calculate Mean Absolute Scaled Error (MASE) and add it to the
plots.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90%
confidence intervals. By default (CI = 0.95), the 95% confidence
intervals are estimated.}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}

\item{xlim}{Limits for x-axis.}

\item{ylim}{Limits for y-axis.}

\item{xlab}{Label for x-axis, "Year" by default.}

\item{ylab}{Label for y-axis, "Index" by default.}

\item{mfrow}{A vector of the form ‘c(nr, nc)’. Subsequent figures will be
drawn in an ‘nr’-by-‘nc’ array on the device by _rows_. Default: NULL.}

\item{mar}{A numerical vector of the form ‘c(bottom, left, top, right)’ which
gives the number of lines of margin to be specified on the four sides of
the plot. The default is ‘c(2, 2, 3, 1) + 0.1’.}

\item{oma}{A vector of the form ‘c(bottom, left, top, right)’ giving the size
of the outer margins in lines of text.}

\item{legend.title}{Legend title. By default no title is used: "".}

\item{legend.pos}{Legend position (default: "topright"). If NULL or NA, no
legend is plotted.}

\item{legend.ncol}{Legend number of columns (default: 1).}

\item{asp}{positive number; the target aspect ratio (columns / rows) in the
graphical output. Default: 2.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
MASE or Invisible \code{NULL}
}
\description{
Hindcasting plot for indices
}
\details{
This function plots the results of the hindcasting cross validation
    analysis for each index.
}
\examples{
data(pol)
inp <- pol$albacore
rep <- fit.spict(inp)
rep <- hindcast(rep, npeels = 5)
plotspict.hindcast(rep)

}
\references{
Carvalho, F., Winker, H., Courtney, D., Kapur, M., Kell, L., Cardinale, M.,
Schirripa, M., Kitakado, T., Yemane, D., Piner, K.R. Maunder, M.N., Taylor,
I., Wetzel, C.R., Doering, K., Johnsonm, K.F., Methot, R. D. (2021). A
cookbook for using model diagnostics in integrated stock assessments.
Fisheries Research, 240, 105959.

Kell, L. T., Kimoto, A., & Kitakado, T. (2016). Evaluation of the prediction
skill of stock assessment using hindcasting. Fisheries research, 183,
119-127.

Kell, L. T., Sharma, R., Kitakado, T., Winker, H., Mosqueira, I., Cardinale,
M., & Fu, D. (2021). Validation of stock assessment methods: is it me or my
model talking?. ICES Journal of Marine Science, 78(6), 2244-2255.

Winker, H., Carvalho, F., & Kapur, M. (2018). JABBA: just another Bayesian
biomass assessment. Fisheries Research, 204, 275-288.
}
\seealso{
\code{\link{hindcast}}
}
