% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.inflsum}
\alias{plotspict.inflsum}
\title{Plots summary of influence statistics of observations.}
\usage{
plotspict.inflsum(rep, stamp = get.version())
}
\arguments{
\item{rep}{A valid result from calc.influence().}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\description{
Plots summary of influence statistics of observations.
}
\details{
TBA
}
