% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.likprof}
\alias{plotspict.likprof}
\title{Plots result of likelihood profiling.}
\usage{
plotspict.likprof(input, logpar = FALSE, stamp = get.version())
}
\arguments{
\item{input}{Result of running likprof.spict().}

\item{logpar}{If TRUE log of parameters are shown.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing but shows a plot.
}
\description{
Plots result of likelihood profiling.
}
\details{
TBA
}
