% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.osar}
\alias{plotspict.osar}
\title{Plot one-step-ahead residuals}
\usage{
plotspict.osar(rep, collapse.I = TRUE, qlegend = TRUE, add.loess = FALSE)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{collapse.I}{Collapse index residuals into one plot. Default: TRUE.}

\item{qlegend}{Plot legend for quarters.}

\item{add.loess}{Add Loess line to residuals}
}
\value{
Nothing.
}
\description{
Plot one-step-ahead residuals
}
\details{
Plots observed versus predicted catches.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
rep <- calc.osa.resid(rep)
plotspict.osar(rep)
}
