% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.retro}
\alias{plotspict.retro}
\title{Plot results of retrospective analysis}
\usage{
plotspict.retro(rep)
}
\arguments{
\item{rep}{A valid result from fit.spict.}
}
\value{
Nothing
}
\description{
Plot results of retrospective analysis
}

