% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.season}
\alias{plotspict.season}
\title{Plot the mean F cycle}
\usage{
plotspict.season(rep, stamp = get.version(), CI = 0.95)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stamp}{Stamp plot with this character string.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}
}
\value{
Nothing.
}
\description{
Plot the mean F cycle
}
\details{
If seasonal data are available the seasonal cycle in the fishing mortality can be estimated. This function plots this mean F cycle.
}
