% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.season}
\alias{plotspict.season}
\title{Plot the mean F cycle}
\usage{
plotspict.season(rep, stamp = get.version())
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stamp}{Stamp plot with this character string.}
}
\value{
Nothing.
}
\description{
Plot the mean F cycle
}
\details{
If seasonal data are available the seasonal cycle in the fishing mortality can be estimated. This function plots this mean F cycle.
}
