% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotspict.tc}
\alias{plotspict.tc}
\title{Plot time constant.}
\usage{
plotspict.tc(rep, main = "Time to Bmsy", stamp = get.version(), CI = 0.95)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{main}{Title of plot.}

\item{stamp}{Stamp plot with this character string.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}
}
\value{
Nothing.
}
\description{
Plot time constant.
}
\details{
Plots the time required for the biomass to reach a certain proportion of Bmsy. The time required to reach 95\% of Bmsy is highlighted.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
plotspict.tc(rep)

}
