% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{pred.catch}
\alias{pred.catch}
\title{Predict the catch of the prediction interval specified in inp}
\usage{
pred.catch(repin, fmsyfac = 1, get.sd = FALSE, exp = FALSE, dbg = 0)
}
\arguments{
\item{fmsyfac}{Projection are made using F = fmsyfac * Fmsy.}

\item{get.sd}{Get uncertainty of the predicted catch.}

\item{exp}{If TRUE report exp of log predicted catch.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}

\item{rep}{Result list as output from fit.spict().}
}
\value{
A vector containing predicted catch (possibly with uncertainty).
}
\description{
Predict the catch of the prediction interval specified in inp
}
