% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.spictcls}
\alias{print.spictcls}
\title{Output a summary of a fit.spict() run.}
\usage{
\method{print}{spictcls}(x, ...)
}
\arguments{
\item{x}{A result report as generated by running fit.spict.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Nothing.
}
\description{
Output a summary of a fit.spict() run.
}
