% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{probdev}
\alias{probdev}
\title{Get probability to be above certain biomass threshold for
    set fishing mortality}
\usage{
probdev(ffac, repin, bbmsyfrac = 0.3, prob = 0.95, getFrac = FALSE,
  verbose = FALSE)
}
\arguments{
\item{ffac}{Fishing mortality factor}

\item{repin}{Result list as output from fit.spict().}

\item{bbmsyfrac}{Fraction of B/Bmsy which is defined as threshold
(Blim = 0.3 Bmsy, Btrigger = 0.5 Bmsy)}

\item{prob}{Probability of being above threshold (bbmsyfrac)}

\item{getFrac}{logical; return}

\item{verbose}{logical; print realised probability, set fishing
mortality factor and deviation}
}
\value{
Returns deviation between targeted and realised probability
    of being above certain biomass threshold under set fishing
    mortality
}
