% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{probdev}
\alias{probdev}
\title{Get probability to be above certain biomass threshold for
    set fishing mortality}
\usage{
probdev(ffac, rep, bfrac = 0.3, prob = 0.95, quant = "logBpBmsynotS",
  reportmode = 2, getFrac = FALSE, verbose = FALSE)
}
\arguments{
\item{ffac}{Fishing mortality factor}

\item{rep}{Result list from fit.spict().}

\item{bfrac}{Fraction of biomass relativ to biomass reference
levels (dependent on \code{quant}), e.g.  fraction of B/Bmsy
which is defined as threshold (Blim = 0.3 Bmsy, Btrigger = 0.5
Bmsy) or fraction of Bp/Bl}

\item{prob}{Probability of being above threshold}

\item{quant}{Quantity to use (options: "logBpBmsy", logBpBl")}

\item{reportmode}{Default 2}

\item{getFrac}{logical; return}

\item{verbose}{logical; print realised probability, set fishing
mortality factor and deviation}
}
\value{
Returns deviation between targeted and realised probability
    of being above certain biomass threshold under set fishing
    mortality
}
\description{
Get probability to be above certain biomass threshold for
    set fishing mortality
}
