% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{probdev}
\alias{probdev}
\title{Get the fishing mortality to be below certain biomass threshold}
\usage{
probdev(ffac, repin, bbmsyfrac = 0.5, prob = 0.95, getFrac = FALSE,
  verbose = FALSE)
}
\arguments{
\item{repin}{Result list as output from fit.spict().}

\item{bbmsyfrac}{Fraction of B/Bmsy which is defined as threshold}

\item{prob}{Probability to above threshold (bbmsyfrac)}
}
\value{
g
}
