% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{shorten.inp}
\alias{shorten.inp}
\title{Shorten time series of input data to specified range}
\usage{
shorten.inp(inp, mintime = NULL, maxtime = NULL)
}
\arguments{
\item{inp}{An input list containing data.}

\item{mintime}{Starting time. If \code{NULL} (default), keep from the start of the time series.}

\item{maxtime}{Ending time. If \code{NULL} (default), keep until the end of the time series.}
}
\value{
List of shortened input time series and input variables as it is returned by \code{\link{check.inp}}
}
\description{
Shorten time series of input data to specified range
}
\details{
Time is given in decimal notation (e.g. 2005.3). If both \code{start} and \code{end}
are \code{NULL}, \code{inp} is returned after running \code{check.inp}.

If code{maxtime} is not set, i.e. the first part of the time series is cut, then management settings are not changed. Otherwise default values are used (see \code{\link{check.inp}}). The \code{ir} vector, which sets up different regimes, is always overwritten.
}
\examples{
inp <- pol$albacore

## Keep only years from 1973 onwards
shorten.inp(inp, mintime = 1973)

## Keep years until 1985
shorten.inp(inp, maxtime = 1985)
\dontrun{
## Empty data set gives an error
shorten.inp(inp, mintime = 1910, maxtime = 1930)
}
}
\seealso{
\code{\link{check.inp}}
}
\author{
T.K. Mildenberger <t.k.mildenberger@gmail.com>
}
