% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stabilityClause}
\alias{stabilityClause}
\title{Apply stability Clause to F or F multiplier}
\usage{
stabilityClause(ffac, lo = 0.8, up = 1.2)
}
\arguments{
\item{ffac}{F factor, multiplier or F to apply stability clause to}

\item{lo}{lower boundary for stability clause (default 0.8)}

\item{up}{upper boundary for stability clause (default 1.2)}
}
\value{
F factor after stability clause was applied
}
