% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spictcls}
\alias{summary.spictcls}
\title{Output a summary of a fit.spict() run.}
\usage{
\method{summary}{spictcls}(object, numdigits = 8)
}
\arguments{
\item{object}{A result report as generated by running fit.spict.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
Nothing. Prints a summary to the screen.
}
\description{
Output a summary of a fit.spict() run.
}
\details{
The output includes the parameter estimates with 95% confidence intervals, estimates of derived parameters (Bmsy, Fmsy, MSY) with 95% confidence intervals, and predictions of biomass, fishing mortality, and catch.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
summary(rep)
}

