% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spictcls}
\alias{summary.spictcls}
\title{Output a summary of a fit.spict() run.}
\usage{
\method{summary}{spictcls}(object, CI = 0.95, ...)
}
\arguments{
\item{object}{A result report as generated by running fit.spict.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Nothing. Prints a summary to the screen.
}
\description{
Output a summary of a fit.spict() run.
}
\details{
The output includes the parameter estimates with 95\% confidence
    intervals, estimates of derived parameters (Bmsy, Fmsy, MSY) with 95\%
    confidence intervals, and predictions of biomass, fishing mortality, and
    catch.
}
\examples{
data(pol)
rep <- fit.spict(pol$albacore)
summary(rep)

}
