% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{sumspict.manage}
\alias{sumspict.manage}
\alias{mansummary}
\title{Print management summary.}
\usage{
sumspict.manage(
  rep,
  include.EBinf = FALSE,
  include.unc = FALSE,
  include.abs = FALSE,
  timeline = TRUE,
  verbose = TRUE
)

mansummary(
  repin,
  include.EBinf = FALSE,
  include.unc = FALSE,
  include.abs = FALSE,
  timeline = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{rep, repin}{A result object as generated by running \code{manage} or \code{add.man.scenario}.}

\item{include.EBinf}{Include EBinf/Bmsy in the output.}

\item{include.unc}{Include uncertainty of management quantities.}

\item{include.abs}{Include absolute B and F in management summary.}

\item{timeline}{(default: FALSE)}

\item{verbose}{Should detailed outputs be provided (default: TRUE).}
}
\value{
List with data frame containing management summary ('est') and data
    frame containing uncertainty of management quantities ('unc') if
    \code{include.unc = TRUE}.
}
\description{
Print management summary.
}
\examples{
\dontrun{
data(pol)
rep <- fit.spict(pol$albacore)
rep <- manage(rep, c(2,4,8))
sumspict.manage(rep)
}

}
