% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{sumspict.manage}
\alias{sumspict.manage}
\title{Print management summary.}
\usage{
sumspict.manage(
  rep,
  include.EBinf = FALSE,
  include.unc = TRUE,
  timeline = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{rep}{A result report as generated by running \code{manage} or \code{man.scenario}.}

\item{include.EBinf}{Include EBinf/Bmsy in the output.}

\item{include.unc}{Include uncertainty of management quantities.}

\item{timeline}{(default: FALSE)}

\item{verbose}{Print more details on observed and predicted time intervals.}
}
\value{
Data frame containing management summary.
}
\description{
Print management summary.
}
