% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.parest}
\alias{sumspict.parest}
\title{Parameter estimates of a fit.spict() run.}
\usage{
sumspict.parest(rep, ndigits = 8, CI = 0.95)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{ndigits}{Present values with this number of digits after the dot.}

\item{CI}{Confidence intervals to be calculated, e.g. 0.9 for the 90\%
confidence intervals. By default (CI = 0.95), the 95\% confidence
intervals are estimated.}
}
\value{
data.frame containing parameter estimates.
}
\description{
Parameter estimates of a fit.spict() run.
}
