% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.predictions}
\alias{sumspict.predictions}
\title{Predictions of a fit.spict() run.}
\usage{
sumspict.predictions(rep, ndigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{ndigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing predictions.
}
\description{
Predictions of a fit.spict() run.
}
