% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.srefpoints}
\alias{sumspict.srefpoints}
\title{Stochastic reference points of a fit.spict() run.}
\usage{
sumspict.srefpoints(rep, stock = 1, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{stock}{Number of the stock whose parameters to present.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing stochastic reference points.
}

