% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sumspict.states}
\alias{sumspict.states}
\title{State estimates of a fit.spict() run.}
\usage{
sumspict.states(rep, ndigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.spict.}

\item{ndigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing state estimates.
}
\description{
State estimates of a fit.spict() run.
}
