% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{take.c}
\alias{take.c}
\title{Calculate management when taking a constant catch (proxy for setting a TAC).}
\usage{
take.c(catch, inpin, repin, maninds, dbg = 0)
}
\arguments{
\item{catch}{Annual catch to take in the prediction period.}

\item{inpin}{Input list.}

\item{repin}{Results list.}

\item{maninds}{Indices of time vector for which to apply management.}

\item{dbg}{Debug flag, dbg=1 some output, dbg=2 more ourput.}
}
\value{
List containing results of management calculations.
}

