% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{txt.stamp}
\alias{txt.stamp}
\title{Add a charater string in the bottom right corner of a plot}
\usage{
txt.stamp(string = get.version(), cex = 0.5, do.flag = NULL)
}
\arguments{
\item{string}{Character string to stamp. If it is an empty string, NULL, or NA the function does nothing}

\item{cex}{Stamp cex.}

\item{do.flag}{If NULL stamp will be added if not in a multi plot, i.e. mean(par()$mfrow) > 1}
}
\value{
Nothing
}
\description{
Add a charater string in the bottom right corner of a plot
}
