% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{txt.stamp}
\alias{txt.stamp}
\title{Add spict version to plot}
\usage{
txt.stamp(string = get.version(), cex = 0.5)
}
\arguments{
\item{string}{Character string to stamp.}
}
\value{
Nothing
}
\description{
Add spict version to plot
}

