% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meyermillar.R
\name{write.bug.file}
\alias{write.bug.file}
\title{Write the BUGS code to a text file}
\usage{
write.bug.file(fn = "tmp.bug", priors)
}
\arguments{
\item{fn}{Filename of to put BUGS code in.}

\item{priors}{List of priors, typically coming from inp$meyermillar$priors.}
}
\value{
Nothing.
}

